% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parameter_from_paper.R
\name{add_parameter_from_paper}
\alias{add_parameter_from_paper}
\title{add_parameter_from_paper}
\usage{
add_parameter_from_paper(
  value,
  parameter_name,
  paper,
  authors,
  release_date,
  journal,
  abstract,
  storage_location
)
}
\description{
add_parameter_from_paper
}
