% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_write.R
\name{add_write}
\alias{add_write}
\title{add_write}
\usage{
add_write(
  path,
  data_product,
  description,
  version,
  file_type,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  use_public
)
}
\arguments{
\item{path}{user-written config.yaml file path}

\item{data_product}{\verb{data_product:} field}

\item{description}{\verb{component:} field}

\item{version}{(optional) \verb{version:} field}

\item{file_type}{(optional) \verb{file_type:} field}

\item{use_data_product}{(optional) \verb{use: data_product:} field}

\item{use_component}{(optional) \verb{use: component:} field}

\item{use_version}{(optional) \verb{use: version:} field}

\item{use_namespace}{(optional) \verb{use: namespace:} field}

\item{use_public}{(optional) \verb{use: public:} field}
}
\description{
Add data product to \code{read} block of user-written config file. Used in
combination with \code{create_config()} for unit testing.
}
\keyword{internal}
