% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_issue.R
\name{attach_issue}
\alias{attach_issue}
\title{attach_issue}
\usage{
attach_issue(
  description,
  severity,
  external_object_doi,
  namespace,
  data_product,
  component,
  key
)
}
\arguments{
\item{description}{a \code{string} containing a free text description of the
issue}

\item{severity}{a \code{numeric} object specifying the severity of the issue}

\item{external_object_doi}{a \code{string} specifying the external_object_doi
(used when issue is to be attached to an external object)}

\item{namespace}{a \code{string} specifying the namespace}

\item{data_product}{a \code{string} specifying the name of the  data_product}

\item{component}{a \code{string} specifying the name of the object_component
component}

\item{key}{API token from data.scrc.uk}
}
\description{
This function is used to attach an issue to either
\itemize{
\item an external object (in which case you must specify the
\code{external_object_doi} argument)
\item a data product (in which case you must specify the \code{namespace} and
\code{data_product} arguments), or
\item a component within a data product (in which case you must specify the
\code{namespace}, \code{data_product}, and \code{component} arguments)
}
}
\examples{
\dontrun{
\donttest{
attach_issue(description = "Data dump caused a spike on the 15th of June",
severity = 19,
external_object_doi = "scottish coronavirus-covid-19-management-information",
namespace = "SCRC",
data_product = "records/SARS-CoV-2/scotland/cases_and_management",
component = "testing_location/date-cumulative",
key = key)
}}

}
