% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ages.R
\name{bin_ages}
\alias{bin_ages}
\title{bin_ages}
\usage{
bin_ages(dat, ageclasses)
}
\arguments{
\item{dat}{dat}

\item{ageclasses}{ageclasses}
}
\description{
bin_ages
}
\examples{
\dontrun{
\dontcheck{
dat <- data.frame(matrix(sample(1000), nrow = 10))
rownames(dat) <- paste0("area", 1:10)
colnames(dat) <- 1:100
bin_ages(dat, "total")
}}

}
