% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataproduct_exists.R
\name{check_dataproduct_exists}
\alias{check_dataproduct_exists}
\title{check_dataproduct_exists}
\usage{
check_dataproduct_exists(
  write_dataproduct,
  write_version,
  write_namespace_id,
  endpoint
)
}
\arguments{
\item{write_dataproduct}{write_dataproduct}

\item{write_version}{write_version}

\item{write_namespace_id}{write_namespace_id}

\item{endpoint}{endpoint}
}
\description{
If a data product already exists with the same name, version, and
namespace, throw an error
}
