% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datetime.R
\name{check_datetime}
\alias{check_datetime}
\title{check_datetime}
\usage{
check_datetime(table, this_field, query_class, this_query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{this_field}{a \code{string} specifying the name of the field}

\item{query_class}{a \code{string} specifying the class of the field}

\item{this_query}{a \code{string} specifying the contents of the field}
}
\description{
check_datetime
}
