% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_entry.R
\name{check_entry}
\alias{check_entry}
\title{check_entry}
\usage{
check_entry(entry, table, field, key, convert_urls = FALSE)
}
\arguments{
\item{entry}{entry}

\item{table}{table}

\item{key}{key}

\item{convert_urls}{convert_urls}
}
\description{
check_entry
}
