% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_field.R
\name{check_field}
\alias{check_field}
\title{check_field}
\usage{
check_field(table, this_field, query_class, this_query, method, endpoint)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{this_field}{a \code{string} specifying the name of the field}

\item{query_class}{a \code{string} specifying the class of the field}

\item{this_query}{a \code{string} specifying the contents of the field}

\item{method}{a \code{string} specifying the method, c("GET", "POST")}

\item{endpoint}{endpoint}
}
\description{
check_field
}
