% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_hdf.R
\name{check_for_hdf5}
\alias{check_for_hdf5}
\title{check_for_hdf5
internal function to check if a component is in a given hdf5 file}
\usage{
check_for_hdf5(filename, component)
}
\arguments{
\item{filename}{file path to the hdf5}

\item{component}{component name to check for}
}
\description{
check_for_hdf5
internal function to check if a component is in a given hdf5 file
}
