% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_handle.R
\name{check_handle}
\alias{check_handle}
\title{check_handle}
\usage{
check_handle(handle, data_product, what, component)
}
\arguments{
\item{handle}{handle}

\item{data_product}{data_product}

\item{what}{c("inputs", "outputs")}

\item{component}{component}
}
\description{
check_handle
}
