% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_local_repo.R
\name{check_local_repo}
\alias{check_local_repo}
\title{check_local_repo}
\usage{
check_local_repo(path)
}
\arguments{
\item{path}{Local repository file path}
}
\value{
\code{boolean}, if local repository is clean (TRUE, else FALSE)
}
\description{
check_local_repo
}
\keyword{internal}
