% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_string.R
\name{check_string}
\alias{check_string}
\title{check_string}
\usage{
check_string(table, this_field, this_query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{this_field}{a \code{string} specifying the name of the field}

\item{this_query}{a \code{string} specifying the contents of the field}
}
\description{
check_string
}
