% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table_exists.R
\name{check_table_exists}
\alias{check_table_exists}
\title{Check Table Exists}
\usage{
check_table_exists(table)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}
}
\value{
Return \code{TRUE} if a table exists, \code{FALSE} if it doesn't
}
\description{
Check Table Exists
}
\keyword{internal}
