% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_yaml_write.R
\name{check_yaml_write}
\alias{check_yaml_write}
\title{check_yaml_write}
\usage{
check_yaml_write(handle, data_product)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying the name of the data product}
}
\description{
check_yaml_write
}
