% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2grid.R
\name{convert2grid}
\alias{convert2grid}
\title{convert2grid}
\usage{
convert2grid(dat, shapefile, subdivisions)
}
\arguments{
\item{dat}{population data}

\item{shapefile}{shapefile}

\item{subdivisions}{subdivisions}
}
\description{
This script transforms age class population data from census geographies to
a grid based system.
}
