% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_grid.R
\name{convert_to_grid}
\alias{convert_to_grid}
\title{Convert census geographies to grid based system}
\usage{
convert_to_grid(dat, shapefile, subdivisions, conversion.table, grid_size)
}
\arguments{
\item{dat}{population data}

\item{shapefile}{shapefile}

\item{subdivisions}{subdivisions}

\item{conversion.table}{conversion table}

\item{grid_size}{grid size}
}
\description{
This script transforms age class population data from census geographies to
a grid based system.
}
