% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_lower.R
\name{convert_to_lower}
\alias{convert_to_lower}
\title{Convert census geographies to lower resolution}
\usage{
convert_to_lower(dat, convert_to, conversion_table)
}
\arguments{
\item{dat}{data}

\item{convert_to}{c("dz", "ur", "iz", "la", "hb", "mmw", "spc")}

\item{conversion_table}{conversion table}
}
\description{
Convert census geographies to lower resolution
}
