% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{create_config}
\usage{
create_config(init_yaml, path, description, script, force = TRUE)
}
\arguments{
\item{init_yaml}{full path of yaml file generated by \verb{fair init}}

\item{path}{user-written config.yaml file path}

\item{description}{\verb{description:} field}

\item{script}{\verb{script:} field}

\item{force}{if \code{TRUE}, then if config.yaml file already exists, it will be
overwritten}
}
\description{
Generates (user generated) config.yaml files for unit tests. Use
\code{add_read()} and \code{add_write()} functions to add read and write
blocks.
}
\keyword{internal}
