% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_estimate.R
\name{create_estimate}
\alias{create_estimate}
\title{create_estimate}
\usage{
create_estimate(filename, path, value, name)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}

\item{path}{a \code{string} specifying the name of the toml file (and path)}

\item{value}{a \code{numeric} object specifying the value}

\item{name}{a \code{string} descriptor}
}
\description{
Function to populate toml file with number type data.
}
\examples{
filename <- "test.toml"
path <- "data-raw/latency"
value <- 2.0
name <- "latency"
create_estimate(filename, path, value, name)

}
