% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_number.R
\name{create_number}
\alias{create_number}
\title{create_number}
\usage{
create_number(h5filename, component, number, name)
}
\arguments{
\item{h5filename}{a \code{string} specifying the name of the hdf5 file}

\item{component}{a \code{string} specifying a location within the hdf5 file}

\item{number}{a \code{numeric} object specifying the value}

\item{name}{a \code{string} descriptor}
}
\description{
Function to populate hdf5 file with number type data.
}
