% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create Table-Type H5 File}
\usage{
create_table(filename, path = ".", component, df, row_names, column_units)
}
\arguments{
\item{filename}{a \code{string} specifying the filename, e.g. "0.1.0.h5"}

\item{path}{a \code{string} specifying the directory in which you want to
save the h5 file; this will be automatically generated if it doesn't
already exist}

\item{component}{a \code{string} specifying a location within the hdf5 file,
e.g. "location/per_week/all_deaths"}

\item{df}{a \code{dataframe} containing the data}

\item{row_names}{(optional) a \code{vector} of rownames}

\item{column_units}{(optional) a \code{vector} comprising column units}
}
\description{
Function to populate hdf5 file with array type data.
}
\examples{
df <- data.frame(column_1 = 1:2, column_2 = 3:4)
rownames(df) <- c("informative_rowname_1", "informative_rowname_2")
filename <- "test_table.h5"

create_table(filename = filename,
             path = ".",
             component = "component_name",
             df = df,
             row_names = rownames(df),
             column_units = c(NA, "m^2"))

file.remove(filename)

}
