% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_number.R
\name{create_version_number}
\alias{create_version_number}
\title{create_version_number}
\usage{
create_version_number(download_date, version)
}
\arguments{
\item{download_date}{download_date}

\item{version}{version}
}
\description{
create_version_number
}
