% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographicData.R
\name{demographicData}
\alias{demographicData}
\title{demographicData}
\usage{
demographicData(
  h5filename = "scrc_demographics.h5",
  datazone_path = "data-raw/population_datazone/sape-2018-persons.xlsx",
  simd_path = "data-raw/DataZone2011_simd2020.csv"
)
}
\arguments{
\item{h5filename}{Name of the hd5f file you want to create}

\item{datazone_path}{path to demographic data file (datazones)}

\item{simd_path}{path to simd data file}
}
\description{
demographicData
}
