% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataproduct.R
\name{download_dataproduct}
\alias{download_dataproduct}
\title{download_dataproduct}
\usage{
download_dataproduct(name, data_dir, version)
}
\arguments{
\item{name}{a \code{string} specifying the name of the data product}

\item{data_dir}{a \code{string} specifying the download directory}

\item{version}{(optional) a \code{string} specifying the version number of
the data product; if version is not specified, the most recent version will
be downloaded}
}
\value{
Returns list comprising two elements
\itemize{
\item{"downloaded_to"}{Where the file was downloaded to}
\item{"components"}{Components contained within the file}
}
}
\description{
download_dataproduct
}
\examples{
\dontrun{
\donttest{
# Automatically download the latest version
download_dataproduct("records/SARS-CoV-2/scotland/cases-and-management/testing", "data-raw")

# Download specified version
download_dataproduct("records/SARS-CoV-2/scotland/cases-and-management/testing", "data-raw", "0.20200920.0")

# Download only version
download_dataproduct("geography/scotland/lookup_table", "data-raw")
}}

}
