% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_external_object.R
\name{download_external_object}
\alias{download_external_object}
\title{Download external object}
\usage{
download_external_object(name, data_dir, version)
}
\arguments{
\item{name}{a \code{string} specifying the name of the external object}

\item{data_dir}{a \code{string} specifying the download directory}

\item{version}{(optional) a \code{string} specifying the version number of
the external object; if version is not specified, the most recent version
will be downloaded}
}
\value{
Returns list comprising two elements
\itemize{
\item{"downloaded_to"}{absolute path of H5 file after downloading}
\item{"components"}{H5 file components}
}
}
\description{
Download external object from the Boydorr server.
}
\examples{
\dontrun{
external_object <- "scottish coronavirus-covid-19-management-information"

# Automatically download the latest version
download_external_object(name = external_object,
                         data_dir = "data-raw")

# Download specified version
download_external_object(name = external_object,
                         data_dir = "data-raw",
                         version = "0.20200920.0")
}

}
\seealso{
Other download functions: 
\code{\link{download_data_product}()},
\code{\link{download_from_database}()},
\code{\link{download_from_url}()}
}
\concept{download functions}
