% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_db.R
\name{download_from_db}
\alias{download_from_db}
\title{download_from_db}
\usage{
download_from_db(url, path, local, filename)
}
\arguments{
\item{url}{url}

\item{path}{path}

\item{local}{local}

\item{filename}{filename}
}
\description{
download_from_db
}
