% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_source_version.R
\name{download_source_version}
\alias{download_source_version}
\title{download_source_version}
\usage{
download_source_version(dataset)
}
\arguments{
\item{dataset}{name of dataset c("scot_dz_shapefile", "scot_dz_demographics",
"scot_gov_deaths", "scot_gov_simd")}
}
\description{
download_source_version
}
