% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2grid.R
\name{dz2grid}
\alias{dz2grid}
\title{dz2grid}
\usage{
dz2grid(dat, datazones, dz_subdivisions)
}
\arguments{
\item{dat}{datazone population data}

\item{datazones}{datazones}

\item{dz_subdivisions}{datazone subdivisions}
}
\description{
This script transforms age class population data from census geographies to
a grid based system.
}
