% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2grid_pop.R
\name{dz2grid_pop}
\alias{dz2grid_pop}
\title{dz2grid_pop}
\usage{
dz2grid_pop(
  population_dz,
  method,
  ageclasses,
  datazones,
  dz_subdivisions,
  postcode
)
}
\arguments{
\item{population_dz}{datazone population data}

\item{method}{c("postcode", "area")}

\item{ageclasses}{vector of class numeric corresponding to the lower bound
of each age class; when missing, a single age class is generated (all
ages combined)}

\item{datazones}{object of class sf and data.frame}

\item{dz_subdivisions}{object of class sf and data.frame}

\item{postcode}{path to postcode shape file}
}
\description{
This script transforms age class population data from census geographies to
a grid based system.
If the census geography is split across two grid cells this script splits
the population between these cells according to the proportion of the
postcodes in the geography which are present in each cell.
i.e. if a datazone containing 5 postcodes is split between 2 grid cells
with 3 postcodes in cell A and 2 in cell B, 60\% of the population goes to cell A.
It takes:
\itemize{
\item shapefile of the census geographies from which the populations should
be drawn
\item a population dataset containing single-year age classes for each
census geography (alternatively a population dataset with preset age
classes can be used) in the structure: 1st column - Geographical identifier,
columns 2:92 single year population totals for age 0 to 89 and 90+ ageclass
\item A grid shapefile which overlays the census geographies. (IF this is not
available this script automatically makes a grid, the dimensions of which
can be chosen (in m))
\item A shapefile containing the boundaries of postcodes contained in the
census geographies
}

If a single-year age class dataset is supplied and a different structure is
desired (i.e. 5-year age classes) the structure of this new age class must
be set as a vector called: age_class_structure
This vector should contain the lower age bound of each age class

There is an issue of rounding when the population are being divided between
cells, I have a work-around which is explained in the code.
}
