% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2grid_simd.R
\name{dz2grid_simd}
\alias{dz2grid_simd}
\title{dz2grid_simd}
\usage{
dz2grid_simd(simd_datazone, gridsize, datazone_sf)
}
\arguments{
\item{simd_datazone}{datazone simd data}

\item{gridsize}{grid size (length) in metres}

\item{datazone_sf}{path to datazone shape file}
}
\description{
Converts simd data from datazone to grid format.
}
