% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2lower.R
\name{dz2lower}
\alias{dz2lower}
\title{dz2lower}
\usage{
dz2lower(dat, convert_to, conversion_table)
}
\arguments{
\item{convert_to}{c("dz", "ur", "iz", "la", "hb", "mmw", "spc")}

\item{conversion_table}{}
}
\description{
dz2lower
}
