% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_init.R
\name{fair_init}
\alias{fair_init}
\title{fair_init}
\usage{
fair_init(
  family_name,
  given_name,
  orcid,
  organisation,
  ror,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{family_name}{a \code{string} specifying family name}

\item{given_name}{a \code{string} specifying given name(s)}

\item{orcid}{a \code{string} specifying an orcid id URL}

\item{organisation}{a \code{string} or vector of \code{strings} specifying
the name of the organisation(s)}

\item{ror}{a \code{string} specifying the organisation identifer (ROR ID) URL}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
fair_init
}
