% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_pull.R
\name{fair_pull}
\alias{fair_pull}
\title{fair_pull}
\usage{
fair_pull(path = "config.yaml", endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{path}{path}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
fair_pull
}
