% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_run.R
\name{fair_run}
\alias{fair_run}
\title{fair_run}
\usage{
fair_run(
  path = "config.yaml",
  endpoint = "http://127.0.0.1:8000/api/",
  skip = FALSE
)
}
\arguments{
\item{path}{string}

\item{endpoint}{a \code{string} specifying the registry endpoint}

\item{skip}{don't bother checking whether the repo is clean}
}
\description{
fair_run
}
