% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-fdp.R
\name{fdp-class}
\alias{fdp-class}
\alias{fdp}
\title{fdp-class}
\description{
fdp-class

fdp-class
}
\details{
Contained for class \code{fdp}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yaml}}{working config.yaml contents}

\item{\code{model_config}}{object URI associated with config.yaml storage location}

\item{\code{submission_script}}{object URI associated with submission script storage}

\item{\code{code_run}}{object URI associated with code run}

\item{\code{inputs}}{metadata associated with code run inputs}

\item{\code{outputs}}{metadata associated with code run outputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fdp$new()}}
\item \href{#method-print}{\code{fdp$print()}}
\item \href{#method-input}{\code{fdp$input()}}
\item \href{#method-output}{\code{fdp$output()}}
\item \href{#method-output_index}{\code{fdp$output_index()}}
\item \href{#method-raise_issue}{\code{fdp$raise_issue()}}
\item \href{#method-finalise_output_hash}{\code{fdp$finalise_output_hash()}}
\item \href{#method-finalise_output_url}{\code{fdp$finalise_output_url()}}
\item \href{#method-clone}{\code{fdp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new fdp object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$new(yaml, model_config, submission_script, code_run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yaml}}{working config.yaml contents}

\item{\code{model_config}}{object URI associated with config.yaml storage location}

\item{\code{submission_script}}{object URI associated with submission script storage location}

\item{\code{code_run}}{object URI associated with code run}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{fdp} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input"></a>}}
\if{latex}{\out{\hypertarget{method-input}{}}}
\subsection{Method \code{input()}}{
Add inputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$input(
  data_product,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  path,
  component_url
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{data_product}

\item{\code{use_data_product}}{use_data_product}

\item{\code{use_component}}{use_component}

\item{\code{use_version}}{use_version}

\item{\code{use_namespace}}{namespace}

\item{\code{path}}{path}

\item{\code{component_url}}{component_url}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output"></a>}}
\if{latex}{\out{\hypertarget{method-output}{}}}
\subsection{Method \code{output()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$output(
  data_product,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  path,
  description
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{data_product}

\item{\code{use_data_product}}{use_data_product}

\item{\code{use_component}}{use_component}

\item{\code{use_version}}{use_version}

\item{\code{use_namespace}}{use_namespace}

\item{\code{path}}{path}

\item{\code{description}}{description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output_index"></a>}}
\if{latex}{\out{\hypertarget{method-output_index}{}}}
\subsection{Method \code{output_index()}}{
Return index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$output_index(data_product, component, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{use_data_product}

\item{\code{component}}{use_component}

\item{\code{version}}{use_version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raise_issue"></a>}}
\if{latex}{\out{\hypertarget{method-raise_issue}{}}}
\subsection{Method \code{raise_issue()}}{
Add issues field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$raise_issue(
  index,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  issue,
  severity
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{text}

\item{\code{use_data_product}}{text}

\item{\code{use_component}}{text}

\item{\code{use_version}}{text}

\item{\code{use_namespace}}{text}

\item{\code{issue}}{text}

\item{\code{severity}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalise_output_hash"></a>}}
\if{latex}{\out{\hypertarget{method-finalise_output_hash}{}}}
\subsection{Method \code{finalise_output_hash()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$finalise_output_hash(use_data_product, use_version, hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_data_product}}{text}

\item{\code{use_version}}{text}

\item{\code{hash}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalise_output_url"></a>}}
\if{latex}{\out{\hypertarget{method-finalise_output_url}{}}}
\subsection{Method \code{finalise_output_url()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$finalise_output_url(
  use_data_product,
  use_component,
  data_product_url,
  component_url
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_data_product}}{text}

\item{\code{use_component}}{text}

\item{\code{data_product_url}}{text}

\item{\code{component_url}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
