% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_resolve_read.R
\name{fdp_resolve_read}
\alias{fdp_resolve_read}
\title{fdp_resolve_read}
\usage{
fdp_resolve_read(this_read, yaml)
}
\arguments{
\item{this_read}{this_read}

\item{yaml}{user written config file}
}
\description{
fdp_resolve_read
}
\keyword{internal}
