% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_resolve_write.R
\name{fdp_resolve_write}
\alias{fdp_resolve_write}
\title{fdp_resolve_write}
\usage{
fdp_resolve_write(this_write, yaml)
}
\arguments{
\item{this_write}{this_write}

\item{yaml}{user written config file}
}
\description{
fdp_resolve_write
}
\keyword{internal}
