% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_run.R
\name{fdp_run}
\alias{fdp_run}
\title{fdp_run}
\usage{
fdp_run(path = "config.yaml", skip = FALSE)
}
\arguments{
\item{path}{string}

\item{skip}{don't bother checking whether the repo is clean}
}
\description{
fdp_run
}
