% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure.R
\name{file_structure}
\alias{file_structure}
\title{file_structure}
\usage{
file_structure(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
Returns the names of the items at the root of the file
}
