% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise.R
\name{finalise}
\alias{finalise}
\title{finalise}
\usage{
finalise(handle, delete_if_empty = FALSE)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{delete_if_empty}{(optional) default is \code{FALSE}}
}
\description{
Push metadata to registry
}
\details{
If a CodeRun does not read an input, write an output, or attach an issue then
\code{delete_if_empty} should delete the CodeRun entry when set to \code{TRUE}.
}
