% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_write_match.R
\name{find_write_match}
\alias{find_write_match}
\title{find_write_match}
\usage{
find_write_match(handle, data_product)
}
\arguments{
\item{handle}{handle}

\item{data_product}{data_product}
}
\description{
Find aliases in working config that match wildcard string
}
