% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_url.R
\name{get_author_url}
\alias{get_author_url}
\title{get_author_url}
\usage{
get_author_url(endpoint)
}
\arguments{
\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
get_author_url
}
\keyword{internal}
