% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components}
\alias{get_components}
\title{get_components}
\usage{
get_components(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
Returns the names of the items at the root of the file
}
