% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataproduct.R
\name{get_dataproduct}
\alias{get_dataproduct}
\title{get_dataproduct}
\usage{
get_dataproduct(
  data_product,
  version,
  namespace,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{data_product}{data_product}

\item{version}{version}

\item{namespace}{namespace}

\item{endpoint}{endpoint}
}
\description{
get_dataproduct
}
