% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity}
\alias{get_entity}
\title{get_entity}
\usage{
get_entity(table, entity_id)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{entity_id}{an \code{integer} specifying the id of an entry}
}
\description{
get_entity
}
\examples{
get_entity("storage_root", 11)

}
