% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry.R
\name{get_entry}
\alias{get_entry}
\title{get_entry}
\usage{
get_entry(table, query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\value{
Returns a \code{list} of fields present in the specified entry
}
\description{
get_entry
}
\examples{
# Get list of entries
data_product <- "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
get_entry("data_product", list(name = data_product))

# Get latest entry
get_entry("data_product")

}
