% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_existing.R
\name{get_existing}
\alias{get_existing}
\title{get_existing}
\usage{
get_existing(table, key, endpoint = "http://data.scrc.uk/api")
}
\arguments{
\item{endpoint}{api endpoint}
}
\description{
get_existing
}
\examples{
get_existing(table = "storage_type")
get_existing(table = "storage_type", key = "ftp")

}
