% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_hash.R
\name{get_file_hash}
\alias{get_file_hash}
\title{Calculate hash from file}
\usage{
get_file_hash(filename)
}
\arguments{
\item{filename}{a \code{string} specifying a filename}
}
\description{
Returns the SHA1 hash of a given file
}
\seealso{
Other get functions: 
\code{\link{get_components}()},
\code{\link{get_entry}()},
\code{\link{get_existing}()},
\code{\link{get_github_hash}()}
}
\concept{get functions}
