% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_hash.R
\name{get_github_hash}
\alias{get_github_hash}
\title{get_github_hash}
\usage{
get_github_hash(repo)
}
\arguments{
\item{repo}{e.g. "ScottishCovidResponse/SCRCdata"}
}
\description{
This function assumes git is installed and located in the System PATH
}
