% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hash.R
\name{get_hash}
\alias{get_hash}
\title{get_hash
Returns the SHA1 hash of a given file}
\usage{
get_hash(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
get_hash
Returns the SHA1 hash of a given file
}
