% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get ID}
\usage{
get_id(table, query = list(), endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
Returns a string or list of strings specifying the URL or URLs of
entries in a table
}
\description{
Retrieve IDs for particular entries or all entries in a table
}
