% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loaded_package_version.R
\name{get_loaded_package_version}
\alias{get_loaded_package_version}
\title{Get loaded package version}
\usage{
get_loaded_package_version(package = "SCRCdataAPI")
}
\arguments{
\item{package}{a \code{string} specifying the package name}
}
\value{
Returns the current package version
}
\description{
Get loaded package version
}
\keyword{internal}
