% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_version.R
\name{get_max_version}
\alias{get_max_version}
\title{get_max_version}
\usage{
get_max_version(data_product, namespace_id)
}
\arguments{
\item{data_product}{data_product}

\item{namespace_id}{namespace_id}
}
\description{
get_max_version
}
