% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_info.R
\name{get_package_info}
\alias{get_package_info}
\title{get_package_info}
\usage{
get_package_info(repo, script_path, package)
}
\arguments{
\item{repo}{\emph{e.g.} "ScottishCovidResponse/SCRCdata"}

\item{script_path}{\emph{e.g.} "scotgov_deaths.R"}

\item{package}{(optional) \emph{e.g.} "SCRCdata"}
}
\value{
Returns a list comprising:
\itemize{
\item the github storage root name: "github"
\item the github repository: e.g. "ScottishCovidResponse/SCRCdata"
\item the github repository version number:
\item the github repo hash
\item the submission script file name
}
}
\description{
Function to retrieve the current version of a GitHub repository and the
GitHub hash associated with it. The repository should contain a DESCRIPTION
file whether it is set up as a package or not. If the repository is set up
as a package, checks are made to ensure remote and local versions of the
package have the same version number.
}
