% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provenance.R
\name{get_provenance}
\alias{get_provenance}
\title{get_provenance}
\usage{
get_provenance(data_product, version, namespace)
}
\arguments{
\item{data_product}{data_product}

\item{version}{version}

\item{namespace}{namespace}
}
\description{
get_provenance
}
