% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remote_package_version.R
\name{get_remote_package_version}
\alias{get_remote_package_version}
\title{Get github Package Version}
\usage{
get_remote_package_version(
  repo = "ScottishCovidResponse/SCRCdataAPI",
  branch = "master"
)
}
\value{
Returns the current package version
}
\description{
Get github Package Version
}
