% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_startup_message.R
\name{get_startup_message}
\alias{get_startup_message}
\title{Get startup message}
\usage{
get_startup_message(repo, package)
}
\arguments{
\item{repo}{repo name}

\item{package}{package name}
}
\value{
returns the startup message
}
\description{
Get startup message
}
