% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage_location.R
\name{get_storage_location}
\alias{get_storage_location}
\title{get_storage_location}
\usage{
get_storage_location(location_id)
}
\arguments{
\item{location_id}{Location ID}
}
\value{
a list of fields from a location_id or null if there was an error
}
\description{
get_storage_location
}
