% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_queryable.R
\name{get_table_queryable}
\alias{get_table_queryable}
\title{Get Queryable Fields}
\usage{
get_table_queryable(table)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}
}
\value{
Returns a character vector of queryable fields
}
\description{
Get Queryable Fields
}
\examples{
get_table_queryable("data_product")

}
\keyword{internal}
