% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_table_readable}
\alias{get_table_readable}
\title{Get Readable Field}
\usage{
get_table_readable(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a character vector of readable fields
}
\description{
Get Readable Field
}
