% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_required.R
\name{get_table_required}
\alias{get_table_required}
\title{Get Required Fields}
\usage{
get_table_required(table, key, type = FALSE)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}

\item{type}{whether to return type of field, returns data frame}
}
\value{
a character vector of required fields
}
\description{
Get Required Fields
}
