% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_tables}
\alias{get_tables}
\title{get tables from registry
use api endpoint to produce a list of tables}
\usage{
get_tables()
}
\value{
a character vector of tables
}
\description{
get tables from registry
use api endpoint to produce a list of tables
}
