% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{get_version}
\usage{
get_version(handle, data_product, namespace)
}
\arguments{
\item{handle}{handle}

\item{data_product}{data_product}

\item{namespace}{namespace}
}
\value{
Returns a character string specifying the version number
}
\description{
get_version
}
