% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_files.R
\name{github_files}
\alias{github_files}
\title{github_files}
\usage{
github_files(repo)
}
\arguments{
\item{repo}{\emph{e.g.}}
}
\description{
github_files
}
\examples{
\donttest{
\dontrun{
filelist <- list_files_in_repo("ScottishCovidResponse/SCRCdata")
files <- grep("scotgov_deaths.R$", filelist, value = TRUE)
}}

}
\keyword{internal}
