% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_intersection.R
\name{grid_intersection}
\alias{grid_intersection}
\title{grid_intersection}
\usage{
grid_intersection(shapefile, gridsize, grid_shp = NA)
}
\arguments{
\item{shapefile}{shapefile}

\item{gridsize}{gridsize}

\item{grid_shp}{gridshape}
}
\description{
grid_intersection
}
