% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_filename.R
\name{increment_filename}
\alias{increment_filename}
\title{increment_filename}
\usage{
increment_filename(scriptdir, filename)
}
\arguments{
\item{scriptdir}{directory}

\item{filename}{file name}
}
\description{
Searches scriptdir for duplicate files and increments filename.
}
