% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_version.R
\name{increment_version}
\alias{increment_version}
\title{increment_version}
\usage{
increment_version(data_product, version)
}
\arguments{
\item{data_product}{a \code{string} specifying the name of the data product}

\item{version}{a \code{string} specifying the version number}
}
\description{
Generate version number by pulling all version numbers from the data registry
and incrementing patch by 1. Otherwise, if a version number is input, checks
are made to determine whether version number is valid.
}
\examples{
increment_version(data_product = "human/demographics/population/scotland")

}
