% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{initialise}
\usage{
initialise(skip)
}
\arguments{
\item{skip}{don't bother checking whether the repo is clean}
}
\description{
Reads in config.yaml file and generates new entry in code_run and returns id
and creates new submission_script in local registry and if necessary creates
a new code_repo entry in local registry.
}
