% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{is_current_version}
\alias{is_current_version}
\title{Check if Package is current version}
\usage{
is_current_version()
}
\value{
Returns true if the loaded package is the same version as the github
package
}
\description{
Check if Package is current version
}
