% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{is_table_writable}
\alias{is_table_writable}
\title{Is table writable}
\usage{
is_table_writable(table)
}
\arguments{
\item{table}{name of table}
}
\value{
boolean if a table is writable
}
\description{
Is table writable
}
