% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_component.R
\name{issue_with_component}
\alias{issue_with_component}
\title{issue_with_component}
\usage{
issue_with_component(
  issue,
  severity,
  data_product,
  namespace,
  component,
  version
)
}
\arguments{
\item{issue}{issue}

\item{severity}{severity}

\item{data_product}{data_product}

\item{namespace}{namespace}

\item{component}{component}

\item{version}{version}
}
\description{
issue_with_component
}
