% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_externalobject.R
\name{issue_with_externalobject}
\alias{issue_with_externalobject}
\title{issue_with_externalobject}
\usage{
issue_with_externalobject(handle, external_object, version, issue, severity)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{external_object}{a \code{string} specifying the name of the external
object}

\item{version}{a \code{string} specifying the version of the external object}

\item{issue}{a \code{string} describing the issue}

\item{severity}{a \code{numeric} value specifying the severity of the issue}
}
\description{
issue_with_externalobject
}
