% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_write.R
\name{link_write}
\alias{link_write}
\title{link_write}
\usage{
link_write(handle, data_product, endpoint)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} representing an external object in the
config.yaml file}

\item{endpoint}{endpoint}
}
\description{
link_write
}
