% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_accessibility.R
\name{new_accessibility}
\alias{new_accessibility}
\title{new_accessibility}
\usage{
new_accessibility(name, description, access_info, key)
}
\arguments{
\item{name}{Name of accessibility type}

\item{description}{Free text description of the accessibility type}

\item{access_info}{Type of accessibility}

\item{key}{GitHub key}
}
\description{
new_accessibility
}
