% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_author.R
\name{new_author}
\alias{new_author}
\title{new_author}
\usage{
new_author(family_name, personal_name, object_id, key)
}
\arguments{
\item{family_name}{e.g.}

\item{personal_name}{e.g.}

\item{object_id}{e.g.}

\item{key}{key}
}
\description{
new_author
}
