% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coderun.R
\name{new_coderun}
\alias{new_coderun}
\title{new_coderun}
\usage{
new_coderun(
  run_date,
  description = "",
  code_repo_id = "",
  model_config = "",
  submission_script_id = "",
  inputs = list(),
  outputs = list(),
  key
)
}
\arguments{
\item{run_date}{the date-time of the code run \emph{e.g.} \code{Sys.time()} or
"2010-07-11 12:15:00 BST"}

\item{description}{(optional) a \code{string} containing a free text
description of the \code{code_run} \emph{e.g.} "Script run to upload and process
scottish coronavirus-covid-19-management-information"}

\item{code_repo_id}{(optional) \emph{e.g.} "https://data.scrc.uk/api/object/154/"}

\item{model_config}{(optional)}

\item{submission_script_id}{\emph{e.g.} "https://data.scrc.uk/api/object/153/"}

\item{inputs}{\emph{e.g.} list("https://data.scrc.uk/api/object_component/875/")}

\item{outputs}{\emph{e.g.} list("https://data.scrc.uk/api/object_component/875/")}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{coderun} table in the data registry
}
