% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_component_issue.R
\name{new_component_issue}
\alias{new_component_issue}
\title{new_component_issue}
\usage{
new_component_issue(component_id, issue_id)
}
\arguments{
\item{issue_id}{}
}
\description{
new_component_issue
}
