% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product.R
\name{new_data_product}
\alias{new_data_product}
\title{Post entry to data_product table}
\usage{
new_data_product(name, version, object_uri, namespace_uri)
}
\arguments{
\item{name}{a \code{string} specifying the name of the \code{data_product}}

\item{version}{a \code{string} specifying the version identifier of the
\code{data_product} (must conform to semantic versioning syntax)}

\item{object_uri}{a \code{string} specifying the URI of the entry in the
\code{object} table}

\item{namespace_uri}{a \code{string} specifying the URI of the entry in the
\code{namespace} table}
}
\description{
Upload information to the \code{data_product} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
