% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product_type.R
\name{new_data_product_type}
\alias{new_data_product_type}
\title{new_data_product_type}
\usage{
new_data_product_type(name, description, key)
}
\arguments{
\item{name}{Name of the data product type}

\item{description}{Free text decription of the data product type}

\item{key}{GitHub key}
}
\description{
new_data_product_type
}
