% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product_version.R
\name{new_data_product_version}
\alias{new_data_product_version}
\title{new_data_product_version}
\usage{
new_data_product_version(
  version_identifier,
  description,
  responsible_person,
  supercedes,
  data_product,
  store,
  accessibility,
  processing_script_version,
  source_versions,
  components,
  model_runs,
  key
)
}
\arguments{
\item{responsible_person}{Name of the responsible person}

\item{supercedes}{Superseded source_version}

\item{store}{Name of the storage_location, e.g. "Model File"}

\item{accessibility}{Name of accessibility type}

\item{model_runs}{}
}
\description{
new_data_product_version
}
