% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product_version_component.R
\name{new_data_product_version_component}
\alias{new_data_product_version_component}
\title{new_data_product_version_component}
\usage{
new_data_product_version_component(
  name,
  responsible_person,
  data_product_version,
  model_runs,
  key
)
}
\arguments{
\item{responsible_person}{Name of the responsible person}

\item{data_product_version}{Name of data product version}

\item{key}{GitHub key}
}
\description{
new_data_product_version_component
}
