% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_external_object.R
\name{new_external_object}
\alias{new_external_object}
\title{new_external_object}
\usage{
new_external_object(
  object_id,
  doi_or_unique_name,
  primary_not_supplement,
  release_date,
  title,
  description,
  source_id,
  original_store_id,
  version
)
}
\arguments{
\item{version}{}
}
\description{
new_external_object
}
