% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file_type.R
\name{new_file_type}
\alias{new_file_type}
\title{Post entry to file_type table}
\usage{
new_file_type(name, extension, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{name}{a \code{string} specifying the name of the file type}

\item{extension}{a \code{string} specifying the filename extension}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{file_type} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
