% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_issue.R
\name{new_issue}
\alias{new_issue}
\title{new_issue}
\usage{
new_issue(severity, description, object_issues, component_issues, key)
}
\arguments{
\item{severity}{e.g.}

\item{description}{e.g.}

\item{object_issues}{e.g.}

\item{component_issues}{e.g.}

\item{key}{key}
}
\description{
new_issue
}
