% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_namespace.R
\name{new_namespace}
\alias{new_namespace}
\title{new_namespace}
\usage{
new_namespace(name, key)
}
\arguments{
\item{name}{a \code{string} specifying the name of the namespace \emph{e.g.} "SCRC"}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{namespace} table in the data registry
}
