% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object.R
\name{new_object}
\alias{new_object}
\title{Post entry to object table}
\usage{
new_object(
  description,
  storage_location_url,
  authors_urls = list(),
  file_type_url
)
}
\arguments{
\item{description}{(optional) a \code{string} containing a free text
description of the \code{object}}

\item{storage_location_url}{(optional) a \code{string} specifying the URL
of an entry in the \code{storage_location} table}

\item{authors_urls}{(optional) a \code{list} of \code{author} URLs to associate
with this \code{object}}

\item{file_type_url}{(optional) a \code{string} specifying the URL
of an entry in the \code{file_type} table}
}
\description{
Upload information to the \code{object} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_author_org}()},
\code{\link{new_object_component}()},
\code{\link{new_organisation}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()}
}
\concept{new functions}
