% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object_author_org.R
\name{new_object_author_org}
\alias{new_object_author_org}
\title{Post entry to object_author_org table}
\usage{
new_object_author_org(
  object_url,
  author_url,
  organisations_urls = list(),
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{object_url}{a \code{string} specifying the URL of an existing
\code{object}}

\item{author_url}{a \code{string} specifying the URL of an existing
\code{author}}

\item{endpoint}{a \code{string} specifying the registry endpoint}

\item{organisations_url}{(optional) a \code{string} specifying the URL of an
existing \code{organisation} to associate with this \code{object_author_org}}
}
\description{
Upload information to the \code{object_author_org} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_organisation}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author_org}()}
}
\concept{new functions}
