% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object_component.R
\name{new_object_component}
\alias{new_object_component}
\title{new_object_component}
\usage{
new_object_component(name, object, key)
}
\arguments{
\item{key}{}
}
\description{
new_object_component
}
