% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_organisation.R
\name{new_organisation}
\alias{new_organisation}
\title{Post entry to organisation table}
\usage{
new_organisation(name, identifier, endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{name}{a \code{string} specifying the name of the \code{organisation}}

\item{identifier}{(optional) a \code{string} representing
the organisation identifier URL, \emph{e.g.} ROR ID}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{organisation} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_author_org}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author_org}()}
}
\concept{new functions}
