% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_processing_script_version.R
\name{new_processing_script_version}
\alias{new_processing_script_version}
\title{new_processing_script_version}
\usage{
new_processing_script_version(
  version_identifier,
  responsible_person,
  supercedes,
  processing_script,
  store,
  accessibility,
  data_product_versions,
  key
)
}
\arguments{
\item{version_identifier}{Version identifier of the source_version, e.g. 1}

\item{responsible_person}{Name of the responsible person}

\item{supercedes}{Superseded source_version}

\item{processing_script}{Name of the processing script}

\item{store}{Name of the storage_location, e.g. "Model File"}

\item{accessibility}{Name of accessibility type}

\item{key}{GitHub key}
}
\description{
new_processing_script_version
}
