% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{new_source}
\usage{
new_source(name, description, responsible_person, store, source_type)
}
\arguments{
\item{name}{name of the source}

\item{description}{free text description of the source}

\item{responsible_person}{reference to the responsible_person}

\item{store}{reference to the storage_location}

\item{source_type}{reference to the source_type}
}
\description{
new_source
}
