% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source_type.R
\name{new_source_type}
\alias{new_source_type}
\title{new_source_type}
\usage{
new_source_type(name, description, key)
}
\arguments{
\item{name}{Name of the source type}

\item{description}{Free text description of the source}

\item{key}{GitHub key}
}
\description{
new_source_type
}
