% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source_version.R
\name{new_source_version}
\alias{new_source_version}
\title{new_source_version}
\usage{
new_source_version(
  version_identifier,
  description,
  responsible_person,
  supercedes,
  source,
  store,
  accessibility,
  key
)
}
\arguments{
\item{version_identifier}{Version identifier of the source_version, e.g. 1}

\item{description}{Free text description of the source_version}

\item{responsible_person}{Name of the responsible person}

\item{supercedes}{Superseded source_version}

\item{source}{Name of the source}

\item{store}{Name of the storage_location, e.g. "Model File"}

\item{accessibility}{Name of accessibility type}

\item{key}{GitHub key}
}
\description{
new_source_version
}
