% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_location.R
\name{new_storage_location}
\alias{new_storage_location}
\title{new_storage_location}
\usage{
new_storage_location(
  name,
  description,
  path,
  hash,
  local_cache_url,
  responsible_person,
  store_root
)
}
\arguments{
\item{store_root}{}
}
\description{
new_storage_location
}
