% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_root.R
\name{new_storage_root}
\alias{new_storage_root}
\title{new_storage_root}
\usage{
new_storage_root(name, root, accessibility, key)
}
\arguments{
\item{name}{e.g.}

\item{root}{e.g.}

\item{accessibility}{e.g.}

\item{key}{key}
}
\description{
new_storage_root
}
