% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_type.R
\name{new_storage_type}
\alias{new_storage_type}
\title{new_storage_type}
\usage{
new_storage_type(name, description, key)
}
\arguments{
\item{name}{Name of storage type, e.g. ftp, https}

\item{description}{Free text description of the storage type, e.g. "File
Transfer Protocol"}

\item{key}{GitHub key}
}
\description{
new_storage_type
}
