% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_text_file.R
\name{new_text_file}
\alias{new_text_file}
\title{Post to text_file table}
\usage{
new_text_file(text, key)
}
\arguments{
\item{text}{a \code{string} containing a free text \emph{i.e.} scripts that are
only a few lines long and don't have a home elsewhere}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{text_file} table in the data registry
}
