% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_user_author.R
\name{new_user_author}
\alias{new_user_author}
\title{Post entry to user_author table}
\usage{
new_user_author(user_url, author_url, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{user_url}{a \code{string} specifying the URL of an existing
\code{user}}

\item{author_url}{a \code{string} specifying the URL of an existing
\code{author}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{user_author} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()}
}
\concept{new functions}
\keyword{internal}
