% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_user_author_org.R
\name{new_user_author_org}
\alias{new_user_author_org}
\title{Post entry to user_author_org table}
\usage{
new_user_author_org(
  user_url,
  author_url,
  organisations_urls = list(),
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{user_url}{a \code{string} specifying the URL of an existing
\code{user}}

\item{author_url}{a \code{string} specifying the URL of an existing
\code{author}}

\item{organisations_urls}{(optional) a \code{list} of URLs specifying which
\code{organisation}s to associate with this \code{user_author_org}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{user_author_org} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_author_org}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_organisation}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()}
}
\concept{new functions}
