% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_data.R
\name{patch_data}
\alias{patch_data}
\title{Patch entry in data registry}
\usage{
patch_data(url, data)
}
\arguments{
\item{url}{url}

\item{data}{data}
}
\description{
Patch entry in data registry
}
\keyword{internal}
