% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_data.R
\name{post_data}
\alias{post_data}
\title{Post entry to data registry}
\usage{
post_data(table, data, ...)
}
\arguments{
\item{table}{table name as a character}

\item{data}{data as a named list}

\item{...}{internal parameters}
}
\description{
Post data to registry
}
\keyword{internal}
