% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_script.R
\name{post_processing_script}
\alias{post_processing_script}
\title{post_processing_script}
\usage{
post_processing_script(
  storage_type,
  storage_root,
  storage_location,
  accessibility,
  processing_script,
  script_version_identifier,
  responsible_person,
  script_version_supercedes,
  key
)
}
\arguments{
\item{storage_type}{Name of storage type, e.g. ftp, https}

\item{storage_root}{Name of the storage root, e.g. "Boydorr"}

\item{storage_location}{Name of the storage_location, e.g. "Model File"}

\item{accessibility}{Name of accessibility type}

\item{responsible_person}{Name of the responsible person}

\item{key}{GitHub key}
}
\description{
post_processing_script
}
