% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_source_data.R
\name{post_source_data}
\alias{post_source_data}
\title{post_source_data}
\usage{
post_source_data(
  storage_type = "ftp",
  storage_root = "Boydorr",
  accessibility = "public",
  source_type = "database",
  source = "statistics.gov.scot",
  storage_location = "deaths-involving-coronavirus-covid-19.csv",
  source_version = "0.1.0",
  target_path = file.path("human", "infection", "SARS-CoV-2", "scotland", "mortality",
    "deaths-involving-coronavirus-covid-19.csv"),
  hash = get_hash(file.path("data-raw", "deaths-involving-coronavirus-covid-19.csv"))
)
}
\arguments{
\item{storage_type}{Name of storage type, e.g. ftp, https}

\item{storage_root}{Name of the storage root, e.g. "Boydorr"}

\item{accessibility}{Name of accessibility type}

\item{source_type}{Name of the source type}

\item{source}{Name of the source}

\item{storage_location}{Name of the storage_location, e.g. "Model File"}

\item{source_version_description}{Free text description of the source_version}

\item{responsible_person}{Name of the responsible person}

\item{key}{GitHub key}
}
\description{
post_source_data
}
