% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scot_dz_demographics.R
\name{process_nrs_demographics}
\alias{process_nrs_demographics}
\title{process_nrs_demographics}
\usage{
process_nrs_demographics(sourcefile, h5filename)
}
\description{
process_nrs_demographics
}
