% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ons_demographics.R
\name{process_ons_demographics}
\alias{process_ons_demographics}
\title{process_ons_demographics}
\usage{
process_ons_demographics(
  sourcefile,
  h5filename,
  datazone_sf,
  grp.names,
  full.names,
  subgrp.names,
  age.classes
)
}
\description{
process_ons_demographics
}
