% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_population.R
\name{process_population}
\alias{process_population}
\title{process_population}
\usage{
process_population(datazone_path)
}
\arguments{
\item{datazone_path}{path to demographic data file (datazones)}
}
\description{
process_population
}
