% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_management.R
\name{process_scotgov_management}
\alias{process_scotgov_management}
\title{process_scotgov_management}
\usage{
process_scotgov_management(sourcefile, h5filename)
}
\description{
process_scotgov_management
}
