% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_simd_income.R
\name{process_scotgov_simd_income}
\alias{process_scotgov_simd_income}
\title{process_scotgov_simd_income}
\usage{
process_scotgov_simd_income(sourcefile, h5filename)
}
\description{
process_scotgov_simd_income
}
