% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scotgov_ur_classification.R
\name{process_scotgov_ur_classification}
\alias{process_scotgov_ur_classification}
\title{process_scotgov_ur_classification}
\usage{
process_scotgov_ur_classification(sourcefile, h5filename)
}
\description{
process_scotgov_ur_classification
}
