% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_shapefiles.R
\name{process_shapefiles}
\alias{process_shapefiles}
\title{process_shapefiles}
\usage{
process_shapefiles(gridsize, datazone_sf, postcode_sf)
}
\arguments{
\item{datazone_sf}{path to datazone shape file}

\item{postcode_sf}{path to postcode shape file}

\item{gridsizes}{vector of lengths (grid sizes), in metres}
}
\description{
process_shapefiles
}
