% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScotDeaths.R
\name{processScotDeaths}
\alias{processScotDeaths}
\title{processScotDeaths}
\usage{
processScotDeaths(h5filename = "deaths-involving-coronavirus-covid-19.h5")
}
\description{
deaths-involving-coronavirus-covid-19
}
