% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue.R
\name{raise_issue}
\alias{raise_issue}
\title{Raise issue with data product object or component}
\usage{
raise_issue(
  index,
  handle,
  component = NA,
  data_product,
  issue,
  severity,
  whole_object = FALSE
)
}
\arguments{
\item{index}{index returned from \verb{link_*()}, \code{read_()}, or \code{write()}}

\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{component}{a \code{string} specifying the component name}

\item{data_product}{a \code{string} specifying the data product name}

\item{issue}{a \code{string} specifying the issue}

\item{severity}{a \code{numeric} value specifying the severity}

\item{whole_object}{a \code{boolean} flag specifying whether or not to
reference the whole_object}
}
\description{
Raise issue with data product object or component
}
\seealso{
Other issue raising functions: 
\code{\link{raise_issue_config}()},
\code{\link{raise_issue_repo}()},
\code{\link{raise_issue_script}()}
}
\concept{issue raising functions}
