% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue_repo.R
\name{raise_issue_repo}
\alias{raise_issue_repo}
\title{raise_issue_repo}
\usage{
raise_issue_repo(handle, issue, severity)
}
\arguments{
\item{handle}{handle}

\item{issue}{issue}

\item{severity}{severity}
}
\description{
raise_issue_repo
}
