% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue_script.R
\name{raise_issue_script}
\alias{raise_issue_script}
\title{raise_issue_script}
\usage{
raise_issue_script(handle, issue, severity)
}
\arguments{
\item{handle}{handle}

\item{issue}{issue}

\item{severity}{severity}
}
\description{
raise_issue_script
}
