% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDataPipeline-package.R
\docType{package}
\name{rDataPipeline-package}
\alias{rDataPipeline-package}
\alias{rDataPipeline}
\title{rDataPipeline}
\description{
FAIR Data Pipeline API
}
\details{
For more information see
\url{https://www.fairdatapipeline.org/}
}
