% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataproduct.R
\name{read_dataproduct}
\alias{read_dataproduct}
\title{read_dataproduct}
\usage{
read_dataproduct(
  path,
  data_product,
  component,
  version,
  use_data_product,
  use_component,
  use_version,
  use_namespace
)
}
\arguments{
\item{path}{config file path}

\item{data_product}{data_product field}

\item{component}{component field}

\item{(optional)}{use_namespace use_namespace field}
}
\description{
read_dataproduct
}
\examples{
\dontrun {
path <- "test_config/config.yaml"

# Write run_metadata block
write_config(path = path,
             description = "test",
             input_namespace = "test_user",
             output_namespace = "test_user")

# Write read block
read_dataproduct(path = path,
                 data_product = "test/array",
                 component = "level/a/s/d/f/s",
                 version = "0.2.0")
}

}
