% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_distribution.R
\name{read_distribution}
\alias{read_distribution}
\title{Read distribution from TOML file}
\usage{
read_distribution(filepath)
}
\arguments{
\item{filepath}{a \code{string} specifying the path and filename of the file
to be read}
}
\description{
Function to read disdtribution type data from toml file.
}
