% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estimate.R
\name{read_estimate}
\alias{read_estimate}
\title{read_estimate}
\usage{
read_estimate(filename)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}
}
\description{
read_estimate
}
