% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_link.R
\name{read_link}
\alias{read_link}
\title{read_link}
\usage{
read_link(handle, alias)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{alias}{a \code{string} representing an external object in the
config.yaml file}
}
\description{
read_link
}
