% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_number.R
\name{read_number}
\alias{read_number}
\title{read_number}
\usage{
read_number(filename)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}
}
\description{
read_number
}
