% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_data_product.R
\name{register_data_product}
\alias{register_data_product}
\title{register_data_product}
\usage{
register_data_product(register_this, datastore, namespace)
}
\arguments{
\item{register_this}{metadata}

\item{datastore}{default local data storage location}

\item{namespace}{}
}
\description{
register_data_product
}
