% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_everything.R
\name{register_everything}
\alias{register_everything}
\title{register_everything}
\usage{
register_everything(
  product_name,
  version_number,
  doi_or_unique_name,
  namespace,
  submission_script,
  original_source_name,
  original_sourceId,
  original_root,
  original_path,
  key
)
}
\arguments{
\item{product_name}{\emph{e.g.}}

\item{version_number}{\emph{e.g.}}

\item{doi_or_unique_name}{\emph{e.g.}}

\item{namespace}{a \code{string} specifying the namespace; by default this
is "SCRC"}

\item{submission_script}{a \code{string} specifying the filename of the
submission script located in \code{inst/namespace/} \emph{e.g.} "nrs_demographics.R"}

\item{original_source_name}{\emph{e.g.}}

\item{original_sourceId}{\emph{e.g.}}

\item{original_root}{\emph{e.g.} "https://www.nrscotland.gov.uk"}

\item{original_path}{\emph{e.g.}}

\item{key}{API token from data.scrc.uk}
}
\description{
This function assumes that (1) the source data is being saved as a
\code{csv} file, (2) the data product is being saved as an \code{h5} file,
(3) the source data and data product are being stored on the boydorr ftp
server (4) the submission script is being stored in the
ScottishCovidResponse/SCRCdata respository
}
