% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_everything_multiple_sources.R
\name{register_everything_multiple_sources}
\alias{register_everything_multiple_sources}
\title{register_everything_multiple_sources}
\usage{
register_everything_multiple_sources(
  product_name,
  version_number,
  doi_or_unique_name,
  namespace,
  submission_script,
  original_source_name,
  original_sourceId,
  original_root,
  original_path,
  accessibility = 0,
  key
)
}
\arguments{
\item{product_name}{\emph{e.g.} "geography/scotland/lookup_table"}

\item{version_number}{\emph{e.g.} "0.1.0"}

\item{doi_or_unique_name}{\emph{e.g.} "Scottish spatial lookup table"}

\item{namespace}{a \code{string} specifying the namespace; by default this
is "SCRC"}

\item{submission_script}{a \code{string} specifying the filename of the
submission script located in \code{inst/namespace/} \emph{e.g.} "scotgov_dz_lookup.R"}

\item{original_source_name}{\emph{e.g.} list(simd = "Scottish Government", dz = "Scottish Government Open Data Repository downloadable file")}

\item{original_sourceId}{\emph{e.g.} list(simd = "https://data.scrc.uk/api/source/3932/", dz = "https://data.scrc.uk/api/source/3976/")}

\item{original_root}{\emph{e.g.} list(simd = "https://www.gov.scot/", dz = "http://statistics.gov.scot/")}

\item{original_path}{\emph{e.g.} list(simd = "binaries/content/documents/govscot/publications/statistics/2020/01/scottish-index-of-multiple-deprivation-2020-data-zone-look-up-file/documents/scottish-index-of-multiple-deprivation-data-zone-look-up/scottish-index-of-multiple-deprivation-data-zone-look-up/govscot\%3Adocument/SIMD\%2B2020v2\%2B-\%2Bdatazone\%2Blookup.xlsx?forceDownload=true", dz = "downloads/file?id=5a9bf61e-7571-45e8-a307-7c1218d5f6b5\%2FDatazone2011Lookup.csv")}

\item{accessibility}{0}

\item{key}{API token from data.scrc.uk}
}
\description{
This function assumes that (1) the source data is being saved as a
\code{csv} file, (2) the data product is being saved as an \code{h5} file,
(3) the source data and data product are being stored on the boydorr ftp
server (4) the submission script is being stored in the
ScottishCovidResponse/SCRCdata respository
}
