% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_external_object.R
\name{register_external_object}
\alias{register_external_object}
\title{register_external_object}
\usage{
register_external_object(yaml, register_this, endpoint)
}
\arguments{
\item{yaml}{config yaml}

\item{register_this}{metadata}

\item{endpoint}{endpoint}
}
\description{
register_external_object
}
