% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_component.R
\name{register_issue_component}
\alias{register_issue_component}
\title{register_issue_component}
\usage{
register_issue_component(handle, this_issue)
}
\arguments{
\item{handle}{handle}

\item{this_issue}{this_issue}
}
\description{
register_issue_component
}
