% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_component.R
\name{register_issue_dataproduct}
\alias{register_issue_dataproduct}
\title{register_issue_dataproduct}
\usage{
register_issue_dataproduct(handle, this_issue)
}
\arguments{
\item{handle}{handle}

\item{this_issue}{this_issue}
}
\description{
register_issue_dataproduct
}
