% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_externalobject.R
\name{register_issue_externalobject}
\alias{register_issue_externalobject}
\title{register_issue_externalobject}
\usage{
register_issue_externalobject(
  description,
  severity,
  external_object_doi,
  version
)
}
\arguments{
\item{description}{description}

\item{severity}{severity}

\item{external_object_doi}{external_object_doi}

\item{version}{version}
}
\description{
register_issue_externalobject
}
