% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_script.R
\name{register_issue_script}
\alias{register_issue_script}
\title{register_issue_script}
\usage{
register_issue_script(handle, this_issue, type)
}
\arguments{
\item{handle}{handle}

\item{this_issue}{this_issue}

\item{type}{type}
}
\description{
register_issue_script
}
