% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_object.R
\name{register_object}
\alias{register_object}
\title{register_object}
\usage{
register_object(register_this, datastore, namespace)
}
\arguments{
\item{register_this}{metadata}

\item{datastore}{default local data storage location}

\item{namespace}{namespace}
}
\description{
register_object
}
