% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_parents.R
\name{remove_empty_parents}
\alias{remove_empty_parents}
\title{remove_empty_parents}
\usage{
remove_empty_parents(path, root)
}
\arguments{
\item{path}{path}

\item{root}{root}
}
\description{
remove_empty_parents
}
