% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_read.R
\name{resolve_read}
\alias{resolve_read}
\title{resolve_read}
\usage{
resolve_read(handle, data_product)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying the name of the data product}
}
\description{
resolve_read
}
