% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_write.R
\name{resolve_write}
\alias{resolve_write}
\title{resolve_data_product}
\usage{
resolve_write(handle, data_product, file_type)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{file_type}{a \code{string} specifying the file type}
}
\description{
resolve_data_product
}
