% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_validation.R
\name{update_validation}
\alias{update_validation}
\title{Private Functions to generate validation cache (inst/validation)}
\usage{
update_validation(key)
}
\description{
Private Functions to generate validation cache (inst/validation)
}
\keyword{internal}
