% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_product.R
\name{upload_data_product}
\alias{upload_data_product}
\title{upload_data_product}
\usage{
upload_data_product(
  storage_root_id,
  path,
  dataset,
  filename,
  version,
  namespace,
  key
)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{path}{e.g.}

\item{dataset}{e.g.}

\item{filename}{e.g.}

\item{version}{e.g.}

\item{namespace}{e.g.}

\item{key}{key}
}
\description{
upload_data_product
}
