% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_external_data.R
\name{upload_external_data}
\alias{upload_external_data}
\title{upload_external_data}
\usage{
upload_external_data(
  source_name,
  source_abbreviation,
  source_website,
  doi_or_unique_name,
  original_source_id,
  original_root_id,
  source_path,
  primary_not_supplement = TRUE,
  local_path,
  storage_root_id,
  target_path,
  download_date,
  version
)
}
\description{
upload_external_data
}
