% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_github_repo.R
\name{upload_github_repo}
\alias{upload_github_repo}
\title{upload_github_repo}
\usage{
upload_github_repo(storage_root_id, repo, hash, version, key)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{hash}{e.g.}

\item{key}{key}

\item{path}{e.g. "ScottishCovidResponse/SCRCdata"}

\item{filename}{e.g.}

\item{run_date}{e.g.}
}
\description{
upload_github_repo
}
