% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_object_links.R
\name{upload_object_links}
\alias{upload_object_links}
\title{Upload object_links metadata to the data registry}
\usage{
upload_object_links(
  run_date,
  description,
  code_repo_id,
  submission_script_id,
  inputs = list(),
  outputs = list(),
  key = key
)
}
\arguments{
\item{run_date}{e.g.}

\item{description}{e.g.}

\item{code_repo_id}{e.g.}

\item{submission_script_id}{e.g.}

\item{inputs}{e.g.}

\item{outputs}{e.g.}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload object_links metadata to the data registry
}
