% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_paper.R
\name{upload_paper}
\alias{upload_paper}
\title{Post paper metadata to the data registry}
\usage{
upload_paper(
  title,
  authors,
  journal,
  journal_abbreviation,
  journal_website,
  release_date,
  abstract,
  keywords,
  doi,
  primary_not_supplement = TRUE,
  version = "1.0.0"
)
}
\arguments{
\item{title}{a \code{string} specifying the title of the paper \emph{e.g.}
"Covid-19: A systemic disease treated with a wide-ranging approach: A case
report"}

\item{authors}{a \code{string} specifying the authors \emph{e.g.} "Massabeti,
Rosanna and Cipriani, Maria Stella and Valenti, Ivana"}

\item{journal}{a \code{string} specifying the full journal name \emph{e.g.}
"Journal of Population Therapeutics and Clinical Pharmacology"}

\item{journal_abbreviation}{a \code{string} specifying the journal
abbreviation \emph{e.g.} "J Popul Ther Clin Pharmacol"}

\item{journal_website}{a \code{string} specifying the journal homepage \emph{e.g.}
"https://www.jptcp.com/index.php/jptcp"}

\item{release_date}{a \code{POSIXct} of format "\%Y-\%m-\%d \%H:\%M:\%S"
specifying the release date of the paper \emph{e.g.}
as.POSIXct("2020-01-01 12:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")}

\item{abstract}{a \code{string} specifying the abstract \emph{e.g.} "At the end
of December 2019, the Health Commission of the"}

\item{keywords}{a \code{string} specifying keywords or keyphrases
seperated by "and", \emph{e.g.} "covid-19 and coronavirus disease and monoclonal
antibodies
\itemize{
\item and non-invasive mechanical ventilation and treatment"
}}

\item{doi}{a \code{string} specifying the doi \emph{e.g.}
"10.15586/jptcp.v27iSP1.691"}

\item{primary_not_supplement}{(optional) an object of class \code{logical}
where \code{TRUE} is the primary source (default) and false is a supplement}

\item{version}{(optional) a \code{string} specifying the version number;
by default this is 1.0.0 (assuming a published paper), set to 0.1.0 for a
preprint}
}
\description{
Post paper metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_github_repo}()},
\code{\link{upload_object_links}()},
\code{\link{upload_source_data}()},
\code{\link{upload_submission_script}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
