% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_processing_script.R
\name{upload_processing_script}
\alias{upload_processing_script}
\title{upload_processing_script}
\usage{
upload_processing_script(
  storage_root_id,
  path,
  repo,
  hash = get_github_hash(repo),
  key
)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{path}{e.g. "ScottishCovidResponse/SCRCdata"}

\item{repo}{repo}

\item{hash}{e.g.}

\item{key}{key}
}
\description{
upload_processing_script
}
