% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_source_data.R
\name{upload_source_data}
\alias{upload_source_data}
\title{upload_source_data}
\usage{
upload_source_data(
  doi_or_unique_name,
  source_id,
  local_path,
  storage_root_id,
  target_path,
  download_date,
  version,
  key
)
}
\arguments{
\item{doi_or_unique_name}{e.g. "scottish deaths-involving-coronavirus-covid-19"}

\item{source_id}{e.g. "Scottish Government Open Data Repository"}

\item{local_path}{e.g. "data-raw/deaths-involving-coronavirus-covid-19.csv"}

\item{storage_root_id}{e.g. "boydorr"}

\item{target_path}{e.g. "human/infection/SARS-CoV-2/scotland/mortality/deaths-involving-coronavirus-covid-19.csv"}

\item{download_date}{e.g. as.POSIXct("2010-07-11 12:15:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")}

\item{version}{e.g. 0}

\item{key}{key}
}
\description{
upload_source_data
}
