% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_submission_script.R
\name{upload_submission_script}
\alias{upload_submission_script}
\title{upload_submission_script}
\usage{
upload_submission_script(storage_root_id, path, hash, run_date, key)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{path}{e.g. "ScottishCovidResponse/SCRCdata"}

\item{hash}{e.g.}

\item{run_date}{e.g.}

\item{key}{key}
}
\description{
upload_submission_script
}
