% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_toml_to_github.R
\name{upload_toml_to_github}
\alias{upload_toml_to_github}
\title{Upload TOML file to GitHub}
\usage{
upload_toml_to_github(
  path_to_toml,
  name = "SCRC",
  email = "scrc@glasgow.ac.uk"
)
}
\arguments{
\item{path_to_toml}{path to the toml file you wish to add to the data repository}

\item{name}{(optional) the name you want to be associated with the git commit}

\item{email}{(optional) the email you want to be associated with the git commit}
}
\description{
Experimental
}
\seealso{
Other upload functions: 
\code{\link{upload_data_product}()},
\code{\link{upload_github_repo}()},
\code{\link{upload_object_links}()},
\code{\link{upload_paper}()},
\code{\link{upload_source_data}()},
\code{\link{upload_submission_script}()}
}
\concept{upload functions}
