% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fields.R
\name{validate_fields}
\alias{validate_fields}
\title{Validate fields}
\usage{
validate_fields(table, data, key, online = TRUE)
}
\arguments{
\item{table}{table name as character}

\item{data}{data as a named list}

\item{key}{API Token as character}

\item{online}{\code{boolean} if FALSE will not check if data registry objects exist}
}
\value{
Returns
}
\description{
Function to validate fields in post data
}
\examples{
\dontrun{
\donttest{
validate_fields("author", list(object = 123, family_name = "Smith",
                               personal_name = "S"), key)
}}

}
\keyword{internal}
