% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table.R
\name{validate_table}
\alias{validate_table}
\title{validate_table}
\usage{
validate_table(table, key)
}
\arguments{
\item{table}{table name as character}

\item{key}{API Token as character}
}
\description{
Function to validate table name
}
