% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_token.R
\name{validate_token}
\alias{validate_token}
\title{validate_token
Function to validate user token}
\usage{
validate_token(key)
}
\arguments{
\item{key}{API Token as character}
}
\description{
validate_token
Function to validate user token
}
\keyword{internal}
