% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config.R
\name{write_config}
\alias{write_config}
\title{write_config}
\usage{
write_config(path, description, input_namespace, output_namespace)
}
\arguments{
\item{output_namespace}{}
}
\description{
Generates (user generated) config.yaml files for unit tests. Use
\code{read_dataproduct()} and \code{write_dataproduct()} functions to
add read and write blocks.
}
\examples{
\dontrun {
write_config("test_config/config.yaml",
             description = "test",
             input_namespace = "test_user",
             output_namespace = "test_user")
}

}
