% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataproduct.R
\name{write_dataproduct}
\alias{write_dataproduct}
\title{write_dataproduct}
\usage{
write_dataproduct(
  path,
  data_product,
  description,
  version,
  use_data_product,
  use_component,
  use_version,
  use_namespace
)
}
\arguments{
\item{path}{config file path}

\item{data_product}{data_product field}

\item{description}{component field}

\item{(optional)}{use_namespace use_namespace field}
}
\description{
write_dataproduct
}
