% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_distribution.R
\name{write_distribution}
\alias{write_distribution}
\title{Write distribution component to TOML file}
\usage{
write_distribution(
  distribution,
  parameters,
  handle,
  data_product,
  component,
  description
)
}
\arguments{
\item{distribution}{a \code{string} specifying the name of the distribution}

\item{parameters}{a \code{list} specifying the distribution parameters}

\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{component}{a \code{string} specifying a location within the toml file}

\item{description}{a \code{string} describing the data product component}
}
\description{
Write distribution component to TOML file
}
\seealso{
Other write functions: 
\code{\link{write_array}()},
\code{\link{write_estimate}()},
\code{\link{write_table}()}
}
\concept{write functions}
