% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_index.R
\name{write_index}
\alias{write_index}
\title{write_index}
\usage{
write_index(index, write, data_product)
}
\arguments{
\item{index}{index}

\item{write}{write}

\item{data_product}{data_product}
}
\description{
write_index
}
