% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_table}
\alias{write_table}
\title{Create table-type H5 file}
\usage{
write_table(df, handle, data_product, component, row_names, column_units)
}
\arguments{
\item{df}{an \code{dataframe} containing the data}

\item{handle}{list}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{component}{a \code{string} specifying a location within the hdf5 file,}

\item{row_names}{(optional) a \code{vector} of rownames}

\item{column_units}{(optional) a \code{vector} comprising column units}
}
\description{
Function to populate hdf5 file with array type data.
}
\seealso{
Other write functions: 
\code{\link{write_array}()}
}
\concept{write functions}
