% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parameter_from_paper.R
\name{add_parameter_from_paper}
\alias{add_parameter_from_paper}
\title{add_parameter_from_paper}
\usage{
add_parameter_from_paper(
  value,
  parameter_name,
  paper,
  authors,
  release_date,
  journal,
  abstract,
  storage_location
)
}
\arguments{
\item{value}{value}

\item{parameter_name}{parameter name}

\item{paper}{paper}

\item{authors}{authors}

\item{release_date}{release date}

\item{journal}{journal}

\item{abstract}{abstract}

\item{storage_location}{storage location}
}
\description{
add_parameter_from_paper
}
