% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataproduct.R
\name{add_write}
\alias{add_write}
\title{add_write}
\usage{
add_write(
  path,
  data_product,
  description,
  version,
  file_type,
  use_data_product,
  use_component,
  use_version,
  use_namespace
)
}
\arguments{
\item{path}{config file path}

\item{data_product}{data_product field}

\item{description}{component field}

\item{version}{(optional) version field}

\item{file_type}{(optional) file type field}

\item{use_data_product}{(optional) use_data_product field}

\item{use_component}{(optional) use_component field}

\item{use_version}{(optional) use_version field}

\item{use_namespace}{(optional) use_namespace field}
}
\description{
add_write
}
