% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_write.R
\name{add_write}
\alias{add_write}
\title{add_write}
\usage{
add_write(
  path,
  data_product,
  description,
  version,
  file_type,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  use_public
)
}
\arguments{
\item{path}{config file path}

\item{data_product}{data_product field}

\item{description}{component field}

\item{version}{(optional) version field}

\item{file_type}{(optional) file type field}

\item{use_data_product}{(optional) use_data_product field}

\item{use_component}{(optional) use_component field}

\item{use_version}{(optional) use_version field}

\item{use_namespace}{(optional) use_namespace field}

\item{use_public}{(optional) use_public field}
}
\description{
Add data product to \code{read} block of user-written config file. Used in
combination with \code{create_config()} for unit testing.
}
\keyword{internal}
