% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_issue.R
\name{attach_issue}
\alias{attach_issue}
\title{attach_issue}
\usage{
attach_issue(
  object_uri,
  object_component = "testing_location/date-cumulative",
  severity,
  description = "Data dump caused a spike on the 15th of June",
  key
)
}
\description{
attach_issue
}
