% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_issue.R
\name{attach_issue}
\alias{attach_issue}
\title{Attach issue to object}
\usage{
attach_issue(description, severity, object, key)
}
\arguments{
\item{description}{a \code{string} containing a free text description of the
\code{issue}}

\item{severity}{an \code{integer} specifying the severity of the \code{issue}}

\item{object}{a \code{list} comprising various elements depending on what
kind of object has an issue; see Details}

\item{key}{API token from data.scrc.uk}
}
\description{
This function is used to attach an issue to an object, which may be an
external object, a data product, or a component within a data product.
}
\details{
The object argument should contain different elements depending on
what kind of object has an issue:
\itemize{
\item{an external object} - {\code{list(external_object_doi = "string",
version = "string")}}
\item{a data product} - {\code{list(data_product = "string",
namespace = "string", version = "string")}}
\item{a component within a data product} - {\code{list(data_product = "string",
namespace = "string", component = "string", version = "string")}}
}
where \code{external_object_doi} is a \code{string} specifying the DOI or
name of the \code{external_object}, \code{namespace} is a \code{string}
specifying the name of the namespace, \code{data_product} is a \code{string}
specifying the name of the \code{data_product}, \code{component} a
\code{string} specifying the name of the \code{object_component}, unique in
the context of \code{object_component} and its \code{object} reference, and
\code{version} is a \code{string} specifying the version number of the
\code{data_product} or \code{external_object}.
}
\examples{
\dontrun{
# Attach issue to an external object
object <- list(external_object_doi = "Scottish spatial lookup table - dz",
               version = "0.1.0")
attach_issue(description = "An issue affecting an external object",
             severity = 10,
             object = object,
             key = key)


# Attach issue to a data product
object <- list(data_product = "records/SARS-CoV-2/scotland/cases-and-management/carehomes",
               namespace = "SCRC",
               version = "0.20200916.0")
attach_issue(description = "An issue affecting a data product",
             severity = 10,
             object = object,
             key = key)

# Attach issue to a component within a data product
object <- list(data_product = "records/SARS-CoV-2/scotland/cases-and-management/carehomes",
               namespace = "SCRC",
               component = "date-country-response_rate",
               version = "0.20200916.0")
attach_issue(description = "An issue affecting a component within a data product",
             severity = 10,
             object = object,
             key = key)
}

}
