% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_issue.R
\name{attach_issue}
\alias{attach_issue}
\title{attach_issue}
\usage{
attach_issue(
  description,
  severity,
  external_object_doi,
  namespace,
  data_product,
  component,
  version,
  key
)
}
\arguments{
\item{description}{a \code{string} containing a free text description of the
\code{issue}}

\item{severity}{an \code{integer} specifying the severity of the \code{issue}}

\item{external_object_doi}{a \code{string} specifying the DOI or name of the
\code{external_object}}

\item{namespace}{a \code{string} specifying the name of the namespace}

\item{data_product}{a \code{string} specifying the name of the
\code{data_product}}

\item{component}{a \code{string} specifying the name of the
\code{object_component}, unique in the context of \code{object_component}
and its \code{object} reference}

\item{version}{a \code{string} specifying the version number of the
data_product or external_object}

\item{key}{API token from data.scrc.uk}
}
\description{
This function is used to attach an issue to an object:
\itemize{
\item{an external object} - {specify the \code{external_object_doi} argument}
\item{a data product} - {specify the \code{namespace} and
\code{data_product} argument}
\item{a component within a data product} - {specify the \code{namespace},
\code{data_product}, and \code{component} argument}
}
See examples.
}
\examples{
\dontrun{
\donttest{
# Attach issue to an external object
attach_issue(description = "An issue affecting an external object",
             severity = 10,
             external_object_doi = "Scottish spatial lookup table - dz",
             version = "0.1.0",
             key = key)

# Attach issue to a data product
attach_issue(description = "An issue affecting a data product",
             severity = 10,
             namespace = "SCRC",
             data_product = "records/SARS-CoV-2/scotland/cases-and-management/carehomes",
             version = "0.20200916.0",
             key = key)

# Attach issue to a component within a data product
attach_issue(description = "An issue affecting a component within a data product",
             severity = 10,
             namespace = "SCRC",
             data_product = "records/SARS-CoV-2/scotland/cases-and-management/carehomes",
             component = "date-country-response_rate",
             version = "0.20200916.0",
             key = key)
}}

}
