% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ages.R
\name{bin_ages}
\alias{bin_ages}
\title{Bin ages}
\usage{
bin_ages(dat, ageclasses)
}
\arguments{
\item{dat}{a \code{data.frame} containing population counts (see Examples
below)}

\item{ageclasses}{a \code{data.frame} containing age class boundaries}
}
\description{
Function used to bin ages in a \code{data.frame}
}
\examples{
# NRS example
df <- matrix(sample(273), ncol = 91)
colnames(df) <- paste0("AGE", 0:90)
colnames(df)[91] <- "AGE90+"
bin_ages(df, c(0, 10, 50, 80))
bin_ages(df, "total")

}
