% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exists.R
\name{check_exists}
\alias{check_exists}
\title{Check if entry exists in the data registry}
\usage{
check_exists(table, query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\value{
Returns \code{TRUE} if the entry exists and \code{FALSE} if it
doesn't
}
\description{
Check whether an entry already exists in a table (in the data registry)
}
