% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_handle.R
\name{check_handle}
\alias{check_handle}
\title{check_handle}
\usage{
check_handle(handle, data_product, what, component)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{what}{element in handle -- one of c("inputs", "outputs")}

\item{component}{a \code{string} specifying the name of the component}
}
\description{
check_handle
}
