% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{check_query}
\alias{check_query}
\title{check query
produces error if table does not exist or if query is not a list}
\usage{
check_query(table, query)
}
\arguments{
\item{table}{name of table}

\item{query}{query to check}
}
\value{
boolean if the query is valid for the table
}
\description{
check query
produces error if table does not exist or if query is not a list
}
