% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_query}
\alias{check_query}
\title{check_query}
\usage{
check_query(table, query)
}
\arguments{
\item{table}{table}

\item{query}{query}
}
\description{
Check whether entries exist in the database
}
