% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_query}
\alias{check_query}
\title{Check if query is valid}
\usage{
check_query(table, query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\value{
Returns \code{TRUE} if the query is valid for the table, \code{FALSE}
if it isn't
}
\description{
Check if query is valid
}
\examples{
check_query("storage_root", list(name = "github"))
check_query("storage_root", list(not_a_field = "github"))
check_query("not_a_table", list(not_a_field = "github"))

}
\keyword{internal}
