% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table_exists.R
\name{check_table_exists}
\alias{check_table_exists}
\title{check table exists}
\usage{
check_table_exists(table)
}
\arguments{
\item{table}{name of table}
}
\value{
boolean if a table exists
}
\description{
check table exists
}
