% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_query.R
\name{clean_query}
\alias{clean_query}
\title{Clean Query}
\usage{
clean_query(data)
}
\arguments{
\item{data}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\description{
Function to clean a query and return it without an api prefix
}
\examples{
data <- list(storage_location = "https://data.scrc.uk/api/storage_location/2")
clean_query(data = data)

}
\keyword{internal}
