% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2grid.R
\name{convert2grid}
\alias{convert2grid}
\title{convert2grid}
\usage{
convert2grid(dat, shapefile, subdivisions, conversion.table, grid_size)
}
\arguments{
\item{dat}{population data}

\item{shapefile}{shapefile}

\item{subdivisions}{subdivisions}

\item{conversion.table}{conversion table}

\item{grid_size}{grid size}
}
\description{
This script transforms age class population data from census geographies to
a grid based system.
}
