% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\title{create_array}
\usage{
create_array(
  filename,
  path,
  component,
  array,
  dimension_names,
  dimension_values,
  dimension_units,
  units
)
}
\arguments{
\item{filename}{a \code{string} specifying the filename, e.g. "0.1.0.h5"}

\item{path}{a \code{string} specifying the directory in which you want to save the h5 file}

\item{component}{a \code{string} specifying a location within the hdf5 file,
e.g. "location/per_week/all_deaths"}

\item{array}{a \code{matrix} containing the data}

\item{dimension_names}{a \code{list} where each element is a vector
containing the labels associated with a particular dimension (e.g.
element 1 corresponds to dimension 1, which corresponds to row names) and
the name of each element describes the contents of each dimension (e.g. age
classes).}

\item{dimension_values}{(optional) a \code{list} of values corresponding to
each dimension (e.g. list element 2 corresponds to columns)}

\item{dimension_units}{(optional) a \code{list} of units corresponding to
each dimension (e.g. list element 2 corresponds to columns)}

\item{units}{(optional) a \code{string} specifying the units of the data as
a whole}
}
\description{
Function to populate hdf5 file with array type data.
}
