% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\title{Create array-type H5 file}
\usage{
create_array(
  filename,
  path = ".",
  component,
  array,
  dimension_names,
  dimension_values,
  dimension_units,
  units
)
}
\arguments{
\item{filename}{a \code{string} specifying the filename, e.g. "0.1.0.h5"}

\item{path}{a \code{string} specifying the directory in which you want to
save the h5 file; this will be automatically generated if it doesn't
already exist}

\item{component}{a \code{string} specifying a location within the hdf5 file,
e.g. "location/per_week/all_deaths"}

\item{array}{an \code{array} containing the data}

\item{dimension_names}{a \code{list} where each element is a vector
containing the labels associated with a particular dimension (e.g.
element 1 corresponds to dimension 1, which corresponds to row names) and
the name of each element describes the contents of each dimension (e.g. age
classes).}

\item{dimension_values}{(optional) a \code{list} of values corresponding to
each dimension (e.g. list element 2 corresponds to columns)}

\item{dimension_units}{(optional) a \code{list} of units corresponding to
each dimension (e.g. list element 2 corresponds to columns)}

\item{units}{(optional) a \code{string} specifying the units of the data as
a whole}
}
\description{
Function to populate hdf5 file with array type data.
}
\examples{
df <- data.frame(a = 1:2, b = 3:4)
rownames(df) <- 1:2
array <- as.matrix(df)

# Create 2-dimensional array
create_array(filename = "test_array_2d.h5",
             path = ".",
             component = "row/column-c1-c2-c3",
             array = array,
             dimension_names = list(rowvalue = rownames(df),
                                    colvalue = colnames(df)))
file.remove("test_array_2d.h5")

# Create 3-dimensional array
create_array(filename = "test_array_3d.h5",
             path = ".",
             component = "row/column/level-c1",
             array = array(c(array, array), dim = c(dim(array), 2)),
             dimension_names = list(rowvalue = rownames(df),
                                    colvalue = colnames(df),
                                    gender = paste0("male", "female")))
file.remove("test_array_3d.h5")

# Create 4-dimensional array
create_array(filename = "test_array_4d.h5",
             path = ".",
             component = "dim1/dim2/dim3/dim4-c1-c2",
             array = array(c(array, array, array), dim = c(dim(array), 2, 2)),
             dimension_names = list(rowvalue = rownames(df),
                                    colvalue = colnames(df),
                                    gender = paste0("male", "female"),
                                    city = paste0("glasgow", "paris")))
file.remove("test_array_4d.h5")

# Create array with values and units
create_array(filename = "test_array_val.h5",
             path = ".",
             component = "row/column-c1-c2",
             array = array,
             dimension_names = list(rowvalue = rownames(df),
                                    colvalue = colnames(df)),
             dimension_values = list(NA, 10),
             dimension_units = list(NA, "km"),
             units = "s")
file.remove("test_array_val.h5")

}
