% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\title{create_array}
\usage{
create_array(
  filename,
  component,
  array,
  dimension_names,
  dimension_values,
  dimension_units
)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the hdf5 file}

\item{component}{a \code{string} specifying a location within the hdf5 file}

\item{array}{a \code{matrix} containing the data}

\item{dimension_names}{a \code{list} where each element is a vector
containing the labels associated with a particular dimension (e.g.
element 1 corresponds to dimension 1, which corresponds to row names) and
the name of each element describes the contents of each dimension (e.g. age
classes).}

\item{dimension_values}{(optional) a \code{list} of values corresponding to
each dimension (e.g. list element 2 corresponds to columns)}

\item{dimension_units}{(optional) a \code{list} of units corresponding to
each dimension (e.g. list element 2 corresponds to columns)}
}
\description{
Function to populate hdf5 file with array type data.
}
\examples{
filename <- "array.h5"
component <- "dz/total"
array <- matrix(1:10, 5)
colnames(array) <- paste0("age", 1:2)
rownames(array) <- paste0("dz", 1:5)
dimension_names <- list(`age classes` = colnames(array),
`area names` = rownames(array))

dimension_values <- list(NA, data.frame(x = 1:2, y = 3:4))
dimension_units <- list(NA, "apples")

create_array(filename, component, array, dimension_names,
dimension_values, dimension_units)

file.h5 <- H5File$new(filename)
file.h5$ls(recursive = TRUE)
file.h5[["dz/total/Dimension_1_names"]][]
file.h5[["dz/total/Dimension_1_title"]][]
file.h5[["dz/total/Dimension_2_names"]][]
file.h5[["dz/total/Dimension_2_title"]][]
file.h5[["dz/total/Dimension_2_units"]][]
file.h5[["dz/total/Dimension_2_values"]][]
file.h5$close_all()

}
