% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{create_config}
\usage{
create_config(path, description, input_namespace, output_namespace)
}
\arguments{
\item{path}{config file path}

\item{description}{description field}

\item{input_namespace}{input_namespace field}

\item{output_namespace}{output_namespace field}
}
\description{
Generates (user generated) config.yaml files for unit tests. Use
\code{read_dataproduct()} and \code{write_dataproduct()} functions to
add read and write blocks.
}
\examples{
\dontrun{
create_config("test_config/config.yaml",
             description = "test",
             input_namespace = "test_user",
             output_namespace = "test_user")
}

}
