% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{create_config}
\usage{
create_config(
  path,
  description,
  input_namespace,
  output_namespace,
  write_data_store = file.path(tempdir(), "datastore", ""),
  force = TRUE,
  local_repo = "local_repo"
)
}
\arguments{
\item{path}{config file path}

\item{description}{description field}

\item{input_namespace}{input_namespace field}

\item{output_namespace}{output_namespace field}

\item{write_data_store}{write_data_store field}

\item{force}{force}

\item{local_repo}{local_repo}
}
\description{
Generates (user generated) config.yaml files for unit tests. Use
\code{add_read()} and \code{add_write()} functions to add read and write
blocks.
}
