% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_distribution}
\alias{create_distribution}
\title{create_distribution}
\usage{
create_distribution(toml_filename, distribution, values, names)
}
\arguments{
\item{toml_filename}{a \code{string} specifying the name of the toml file}

\item{distribution}{a \code{string} specifying the name of the distribution}

\item{values}{a \code{numeric} \code{vector} specifying parameter values}

\item{names}{a \code{vector} specifying the name of each parameter value}
}
\description{
Function to populate toml file with distribution type data.
}
