% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_distribution}
\alias{create_distribution}
\title{create_distribution}
\usage{
create_distribution(filename, path, descriptor, distribution, parameters)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}

\item{path}{a \code{string} specifying the name of the toml file (and path)}

\item{descriptor}{a \code{string} specifying an associated descriptor}

\item{distribution}{a \code{string} specifying the name of the distribution}

\item{parameters}{a named \code{list} specifying parameter values and
associated names (see example)}
}
\description{
Function to populate toml file with distribution type data.
}
\examples{
filename <- "test.toml"
path <- "data-raw/latency"
descriptor <- "latency"
distribution <- "gamma"
parameters <- list(shape = 2.0, scale = 3.0)
create_distribution(filename, path, descriptor, distribution, parameters)

}
