% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_estimate.R
\name{create_estimate}
\alias{create_estimate}
\title{create_estimate}
\usage{
create_estimate(filename, path, parameters)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}

\item{path}{a \code{string} specifying the directory in which you want to
save the toml file; this will be automatically generated if it doesn't
already exist}

\item{parameters}{a \code{list} (see Examples)}
}
\description{
Function to populate toml file with point-estimate type data.
}
\examples{
# Write a single estimate into a toml file
filename <- "test.toml"

create_estimate(filename = filename,
                path = ".",
                parameters = list(asymptomatic-period = 192.0))

file.remove(filename)

# Write multiple estimates into a toml file
filename <- "anothertest.toml"

create_estimate(filename = filename,
                path = ".",
                parameters = list(asymptomatic-period = 192.0,
                                  standard-deviation = 10.2))

file.remove(filename)

}
