% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_estimate.R
\name{create_estimate}
\alias{create_estimate}
\title{create_estimate}
\usage{
create_estimate(filename, path, parameters)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}

\item{path}{a \code{string} specifying the name of the toml file (and path)}

\item{parameters}{}
}
\description{
Function to populate toml file with number type data.
}
\examples{
filename <- "multi.toml"
path <- "data-raw/multi"
parameters <- list(asymptomatic_period = 192.0)
create_estimate(filename, path, parameters)

filename <- "multi.toml"
path <- "data-raw/multi"
parameters <- list(asymptomatic_period = 192.0, latent_period = 123.12)
create_estimate(filename, path, parameters)

}
