% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_estimate.R
\name{create_estimate}
\alias{create_estimate}
\title{Create Estimate-Type TOML File}
\usage{
create_estimate(filename, path, parameters)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file}

\item{path}{a \code{string} specifying the directory in which you want to
save the toml file; this will be automatically generated if it doesn't
already exist}

\item{parameters}{a \code{list} (see Examples)}
}
\description{
Function to populate toml file with point-estimate type data, \emph{e.g.}
\itemize{
\item{Generate a single toml file containing a single point-estimate component}
\item{Generate a single toml file containing multiple point-estimate components}
}
If a file already exists at the specified location, an error will be returned.
}
\examples{
filename <- "0.1.0.toml"
data_product_name <- "some/descriptive/name/asymptomatic-period"

# Write a single estimate into a toml file
create_estimate(filename = filename,
                path = data_product_name,
                parameters = list(`asymptomatic-period` = 192.0))

file.remove(file.path(data_product_name, filename))

# Write multiple estimates into a toml file
create_estimate(filename = filename,
                path = data_product_name,
                parameters = list(`asymptomatic-period-1` = 192.0,
                                  `asymptomatic-period-2` = 190.2))

file.remove(file.path(data_product_name, filename))

}
