% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_number.R
\name{create_number}
\alias{create_number}
\title{create_number}
\usage{
create_number(toml_filename, value, name)
}
\arguments{
\item{toml_filename}{a \code{string} specifying the name of the toml file}

\item{value}{a \code{numeric} object specifying the value}

\item{name}{a \code{string} descriptor}
}
\description{
Function to populate toml file with number type data.
}
