% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_version_number.R
\name{create_version_number}
\alias{create_version_number}
\title{create_version_number}
\usage{
create_version_number(
  download_date = NULL,
  version = NULL,
  major = NULL,
  minor = "1",
  patch = NULL
)
}
\arguments{
\item{download_date}{download_date
This can either be a date or datetime
but must include the full year
\emph{e.g} from
\code{Sys.Date()
2020-01-01}
Or from
\code{Sys.time()
2020-01-01 00:00:00 BST}
note: also accepts / delimited dates
\emph{e.g} \code{01/02/2020}
or \code{2020/01/01}
and accepts date without delimiters but assumes ddmmyyyy or yyyymmdd
\emph{e.g.} \code{20200201}}

\item{version}{version number using major.minor.patch numbering
\emph{e.g.} \code{0.1.0}}

\item{major}{major number if not using \code{version}}

\item{minor}{minor number if not using \code{date}}

\item{patch}{patch number if not using \code{version}}
}
\value{
returns a character vector in the format of \code{YYYYMMDD.major.minor.patch}
\emph{e.g.} \code{20200101.0.1.0}
}
\description{
create_version_number
}
