% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{demographics}
\usage{
demographics(
  filename = "demographics.h5",
  datazone_sf = "data-raw/shapefiles/SG_DataZone_Bdry_2011.shp",
  datazone_path = "data-raw/sape-2018-persons.xlsx",
  grp.names = c("dz", "ur", "iz", "la", "hb", "mmw", "spc", "grid1km", "grid10km"),
  subgrp.names = c("total", "1year", "5year", "10year"),
  age.classes = list("total", 0:90, seq(0, 90, 5), seq(0, 90, 10))
)
}
\description{
Estimated population by sex, single year of age, 2011 Data Zone area, and
council area: 30 June 2018.
}
