% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{demographics}
\usage{
demographics(
  h5filename = "scrc_demographics.h5",
  gridsize = 10,
  ageclasses = seq(0, 90, 5),
  datazone_sf = "data-raw/shapefiles/SG_DataZone_Bdry_2011.shp",
  postcode_sf = "data-raw/shapefiles/PC_Cut_20_1.shp",
  datazone_path = "data-raw/sape-2018-persons.xlsx",
  simd_path = "data-raw/DataZone2011_simd2020.csv"
)
}
\arguments{
\item{h5filename}{Name of the hd5f file you want to create}

\item{gridsize}{grid size (length) in metres}

\item{ageclasses}{vector of class numeric corresponding to the lower bound
of each age class; when missing, a single age class is generated (all
ages combined)}

\item{datazone_sf}{path to datazone shape file}

\item{postcode_sf}{path to postcode shape file}

\item{datazone_path}{path to demographic data file (datazones)}

\item{simd_path}{path to simd data file}
}
\description{
demographics
}
