% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_product.R
\name{download_data_product}
\alias{download_data_product}
\title{Download data product}
\usage{
download_data_product(data_product, version, namespace, data_dir)
}
\arguments{
\item{data_product}{a \code{string} specifying the name of the data product}

\item{version}{(optional) a \code{string} specifying the version number of
the data product; if version is not specified, the most recent version will
be downloaded}

\item{namespace}{namespace}

\item{data_dir}{a \code{string} specifying the download directory}
}
\value{
Returns list comprising two elements
\itemize{
\item{"downloaded_to"}{absolute path of H5 file after downloading}
\item{"components"}{H5 file components}
}
}
\description{
Download data product from the Boydorr server.
}
\seealso{
Other download functions: 
\code{\link{download_external_object}()},
\code{\link{download_from_database}()},
\code{\link{download_from_url}()}
}
\concept{download functions}
