% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_url.R
\name{download_from_url}
\alias{download_from_url}
\title{Download source file from URL}
\usage{
download_from_url(source_root, source_path, path, filename)
}
\arguments{
\item{source_root}{a \code{string} specifying the source root}

\item{source_path}{a \code{string} specifying the source path}

\item{path}{a \code{string} specifying where the file will be saved}

\item{filename}{a \code{string} specifying the filename (the name given to
the saved file)}
}
\description{
This function will download a file from a url
}
\seealso{
Other download functions: 
\code{\link{download_from_database}()}
}
\concept{download functions}
