% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz2grid_simd.R
\name{dz2grid_simd}
\alias{dz2grid_simd}
\title{dz2grid_simd}
\usage{
dz2grid_simd(simd_datazone, datazone_sf, grid_size = 10000)
}
\arguments{
\item{simd_datazone}{datazone simd data}

\item{datazone_sf}{path to datazone shape file}

\item{grid_size}{grid size (length) in metres}
}
\description{
Converts simd data from datazone to grid format.
}
