% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_init.R
\name{fair_init}
\alias{fair_init}
\title{fair_init}
\usage{
fair_init(name, identifier, endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{name}{a \code{string} specifying the full name or organisation name of
the \code{author}; note that at least one of name or identifier must be
specified}

\item{identifier}{(optional) a \code{string} specifying the full URL
identifier (\emph{e.g.} ORCiD or ROR ID) of the \code{author}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
fair_init
}
