% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-fdp.R
\name{fdp-class}
\alias{fdp-class}
\alias{fdp}
\title{fdp-class}
\description{
fdp-class

fdp-class
}
\details{
Contained for class \code{fdp}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yaml}}{text}

\item{\code{model_config}}{text}

\item{\code{submission_script}}{text}

\item{\code{inputs}}{text}

\item{\code{outputs}}{text}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fdp$new()}}
\item \href{#method-input}{\code{fdp$input()}}
\item \href{#method-write_dataproduct}{\code{fdp$write_dataproduct()}}
\item \href{#method-clone}{\code{fdp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new fdp object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$new(yaml, model_config, submission_script)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yaml}}{text}

\item{\code{model_config}}{text}

\item{\code{submission_script}}{text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{fdp} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input"></a>}}
\if{latex}{\out{\hypertarget{method-input}{}}}
\subsection{Method \code{input()}}{
Add inputs field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$input(alias, externalobject_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{text}

\item{\code{externalobject_id}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_dataproduct"></a>}}
\if{latex}{\out{\hypertarget{method-write_dataproduct}{}}}
\subsection{Method \code{write_dataproduct()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$write_dataproduct(data_product, path, component)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{text}

\item{\code{path}}{text}

\item{\code{component}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
