% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-fdp.R
\name{fdp-class}
\alias{fdp-class}
\alias{fdp}
\title{fdp-class}
\description{
fdp-class

fdp-class
}
\details{
Contained for class \code{fdp}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yaml}}{text}

\item{\code{model_config}}{text}

\item{\code{submission_script}}{text}

\item{\code{inputs}}{text}

\item{\code{outputs}}{text}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fdp$new()}}
\item \href{#method-print}{\code{fdp$print()}}
\item \href{#method-input}{\code{fdp$input()}}
\item \href{#method-write_dataproduct}{\code{fdp$write_dataproduct()}}
\item \href{#method-output_index}{\code{fdp$output_index()}}
\item \href{#method-raise_issue_component}{\code{fdp$raise_issue_component()}}
\item \href{#method-write_component_id}{\code{fdp$write_component_id()}}
\item \href{#method-write_dataproduct_id}{\code{fdp$write_dataproduct_id()}}
\item \href{#method-clone}{\code{fdp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new fdp object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$new(yaml, model_config, submission_script)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yaml}}{text}

\item{\code{model_config}}{text}

\item{\code{submission_script}}{text}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{fdp} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input"></a>}}
\if{latex}{\out{\hypertarget{method-input}{}}}
\subsection{Method \code{input()}}{
Add inputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$input(alias, type, doi_or_unique_name, title, version, path, object_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{text}

\item{\code{type}}{text}

\item{\code{doi_or_unique_name}}{text}

\item{\code{title}}{text}

\item{\code{version}}{text}

\item{\code{path}}{text}

\item{\code{object_id}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_dataproduct"></a>}}
\if{latex}{\out{\hypertarget{method-write_dataproduct}{}}}
\subsection{Method \code{write_dataproduct()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$write_dataproduct(data_product, path, component, description, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{text}

\item{\code{path}}{text}

\item{\code{component}}{text}

\item{\code{description}}{text}

\item{\code{version}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output_index"></a>}}
\if{latex}{\out{\hypertarget{method-output_index}{}}}
\subsection{Method \code{output_index()}}{
Return index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$output_index(this_data_product, this_component, this_version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{this_data_product}}{text}

\item{\code{this_component}}{text}

\item{\code{this_version}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raise_issue_component"></a>}}
\if{latex}{\out{\hypertarget{method-raise_issue_component}{}}}
\subsection{Method \code{raise_issue_component()}}{
Add issues field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$raise_issue_component(component, data_product, version, issue, severity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{text}

\item{\code{data_product}}{text}

\item{\code{version}}{text}

\item{\code{issue}}{text}

\item{\code{severity}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_component_id"></a>}}
\if{latex}{\out{\hypertarget{method-write_component_id}{}}}
\subsection{Method \code{write_component_id()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$write_component_id(data_product, component, component_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{text}

\item{\code{component}}{text}

\item{\code{component_id}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_dataproduct_id"></a>}}
\if{latex}{\out{\hypertarget{method-write_dataproduct_id}{}}}
\subsection{Method \code{write_dataproduct_id()}}{
Add outputs field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$write_dataproduct_id(data_product, data_product_id, version, hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_product}}{text}

\item{\code{data_product_id}}{text}

\item{\code{version}}{text}

\item{\code{hash}}{text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
