% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise.R
\name{finalise}
\alias{finalise}
\title{Finalise code run}
\usage{
finalise(handle, delete_if_empty = FALSE, delete_if_duplicate = FALSE)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{delete_if_empty}{(optional) default is \code{FALSE}; see Details}

\item{delete_if_duplicate}{(optional) default is \code{FALSE}; see Details}
}
\description{
Finalise Code Run and push associated metadata to the local registry.
}
\details{
If a Code Run does not read an input, write an output, or attach an issue,
then delete the Code Run entry when \code{delete_if_empty} is set to \code{TRUE}.

If a data product has the same hash as a previous version, remove it from
the registry when \code{delete_if_duplicate} is set to \code{TRUE}.
}
