% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_read_match.R
\name{find_read_match}
\alias{find_read_match}
\title{find_read_match}
\usage{
find_read_match(handle, data_product = "disease/SARS-CoV-2/input/*")
}
\arguments{
\item{handle}{handle}

\item{data_product}{data_product}
}
\description{
find_read_match
}
