% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findme.R
\name{findme}
\alias{findme}
\title{findme}
\usage{
findme(file, endpoint)
}
\arguments{
\item{file}{file path}

\item{endpoint}{endpoint}
}
\description{
Returns metadata associated with the calculated hash of a target file.
When multiple entries exist in the data registry all are returned.
}
