% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components}
\alias{get_components}
\title{Get H5 file components}
\usage{
get_components(filename)
}
\arguments{
\item{filename}{a \code{string} specifying a filename}
}
\value{
Returns the names of the items at the root of the file
}
\description{
Returns the names of the items at the root of the file
}
\seealso{
Other get functions: 
\code{\link{get_entry}()},
\code{\link{get_existing}()},
\code{\link{get_file_hash}()},
\code{\link{get_github_hash}()}
}
\concept{get functions}
