% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity}
\alias{get_entity}
\title{Get entity from url}
\usage{
get_entity(url)
}
\arguments{
\item{url}{a \code{string} specifying the url of an entry}
}
\description{
Get entity from url
}
\keyword{internal}
