% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry.R
\name{get_entry}
\alias{get_entry}
\title{Return all fields associated with a table entry in the data registry}
\usage{
get_entry(table, query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\value{
Returns a \code{list} of fields present in the specified entry
}
\description{
Return all fields associated with a table entry in the data registry
}
\examples{
\dontrun{
# Get list of entries
data_product <- "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
get_entry("storage_location",
list(path = "/Users/SoniaM/datastore/config/20210422-115831.yaml",
hash = "e3caf6a4674bde77e6bd3cbc85b6a867adf7397b",
storage_root = "7"))
}

}
\seealso{
Other get functions: 
\code{\link{get_components}()},
\code{\link{get_existing}()},
\code{\link{get_file_hash}()},
\code{\link{get_github_hash}()},
\code{\link{get_package_info}()}
}
\concept{get functions}
