% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry.R
\name{get_entry}
\alias{get_entry}
\title{get_entry}
\usage{
get_entry(table, query = list())
}
\arguments{
\item{table}{table must exist}

\item{query}{a \code{list} of fields and values to query if no query
is provided it will return the last entry}
}
\description{
get_entry
}
\examples{
get_entry("storage_root", list(name = "github"))

}
