% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry.R
\name{get_entry}
\alias{get_entry}
\title{Return all fields associated with a table entry in the data registry}
\usage{
get_entry(table, query, endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
Returns a \code{list} of fields present in the specified entry
}
\description{
Return all fields associated with a table entry in the data registry
}
\seealso{
Other get functions: 
\code{\link{get_components}()},
\code{\link{get_existing}()},
\code{\link{get_file_hash}()},
\code{\link{get_github_hash}()}
}
\concept{get functions}
