% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_existing.R
\name{get_existing}
\alias{get_existing}
\title{get_existing}
\usage{
get_existing(table, limit_results = TRUE, detail = "all")
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{limit_results}{a \code{boolean} specifying whether or not to limit
the results, default is \code{TRUE}}

\item{detail}{a \code{string} specifying what level of detail to return;
options are \code{"all"} for all details or \code{"id"} for just URL and IDs}
}
\value{
Returns a \code{data.frame} of entries in table, default is limited
to 100 entries
}
\description{
Get entries (from the data registry) in a particular table
}
\examples{
# Get all entries in storage_root
get_existing("storage_root", limit_results = FALSE)

# Get 100 newest entries in data_product
get_existing("data_product")

}
