% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_existing.R
\name{get_existing}
\alias{get_existing}
\title{get_existing}
\usage{
get_existing(endpoint = "http://data.scrc.uk/api", table, key)
}
\arguments{
\item{endpoint}{api endpoint}

\item{user_keys}{}
}
\description{
get_existing
}
