% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get fields from table}
\usage{
get_fields(table, key, filter_fields = "all", type = FALSE)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}

\item{filter_fields}{which fields you want returned \emph{e.g.} all, none,
readable, writable, optional}

\item{type}{logical whether to include types this will return a dataframe}
}
\value{
a character vector of fields or a list of fields if filter fields is
set to "none"
}
\description{
Use API endpoint to produce a list of fields for a table.
}
\details{
Requires API key
}
