% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get fields from table}
\usage{
get_fields(table, key)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}
}
\value{
a dataframe of fields and their attributes
set to "none"
}
\description{
Use API endpoint to produce a list of fields for a table.
}
\details{
Requires API key
}
