% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get fields from table}
\usage{
get_fields(table, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
Returns a \code{data.frame} of fields and their attributes set to
"none"
}
\description{
Use API endpoint to produce a list of fields for a table. Requires API key.
}
\keyword{internal}
