% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_hash.R
\name{get_github_hash}
\alias{get_github_hash}
\title{Get current GitHub hash}
\usage{
get_github_hash(repo)
}
\arguments{
\item{repo}{a \code{string} specifying the github username/repository}
}
\description{
Get the hash of the latest commit in the master branch of a particular
repository. This function assumes git is installed and located in the
System PATH.
}
\seealso{
Other get functions: 
\code{\link{get_components}()},
\code{\link{get_entry}()},
\code{\link{get_existing}()},
\code{\link{get_file_hash}()}
}
\concept{get functions}
