% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h5_from_object_id.R
\name{get_h5_from_object_id}
\alias{get_h5_from_object_id}
\title{get_h5_from_object_id}
\usage{
get_h5_from_object_id(data_dir, obj_id)
}
\arguments{
\item{data_dir}{Data Directory}

\item{obj_id}{object ID}
}
\value{
H5 file path after downloading
}
\description{
get_h5_from_object_id
}
