% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h5_from_object_id.R
\name{get_h5_from_object_id}
\alias{get_h5_from_object_id}
\title{get_h5_from_object_id}
\usage{
get_h5_from_object_id(obj_id, data_dir = "data-raw")
}
\arguments{
\item{obj_id}{an \code{integer} specifying the id of an entry}

\item{data_dir}{a \code{string} specifying the directory in which the file
should be saved}
}
\value{
Returns a \code{list} comprising two elements
\itemize{
\item{downloaded_to}{absolute path of H5 file after downloading}
\item{components}{H5 file components}
}
}
\description{
get_h5_from_object_id
}
