% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_startup_message.R
\name{get_loaded_package_version}
\alias{get_loaded_package_version}
\title{Get loaded Package Version}
\usage{
get_loaded_package_version(package = "SCRCdataAPI")
}
\value{
Returns the current package version
}
\description{
Get loaded Package Version
}
