% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_info.R
\name{get_package_info}
\alias{get_package_info}
\title{Get GitHub package info}
\usage{
get_package_info(repo, script_path, package)
}
\arguments{
\item{repo}{a \code{string} specifying the github username/repository}

\item{script_path}{a \code{string} specifying the submission script filename
and extension}

\item{package}{(optional) a \code{string} specifying the name of the package
containing the submission script}
}
\value{
Returns a list comprising:
\itemize{
\item the github storage root name: "github"
\item the github repository: e.g. "ScottishCovidResponse/SCRCdata"
\item the github repository version number:
\item the github repo hash
\item the submission script file name
}
}
\description{
Function to retrieve the current version of a GitHub repository and the
GitHub hash associated with it. The repository should contain a DESCRIPTION
file whether it is set up as a package or not. If the repository is set up
as a package, checks are made to ensure remote and local versions of the
package have the same version number.
}
\examples{
\donttest{
\dontrun{
get_package_info(repo = "ScottishCovidResponse/SCRCdata",
script_path = "scotgov_deaths.R",
package = "SCRCdata")
}}

}
\seealso{
Other get functions: 
\code{\link{get_components}()},
\code{\link{get_entry}()},
\code{\link{get_existing}()},
\code{\link{get_file_hash}()},
\code{\link{get_github_hash}()}
}
\concept{get functions}
