% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remote_package_version.R
\name{get_remote_package_version}
\alias{get_remote_package_version}
\title{Get Remote (GitHub) Package Version}
\usage{
get_remote_package_version(repo = "ScottishCovidResponse/SCRCdataAPI")
}
\arguments{
\item{repo}{a \code{string} specifying the github username/repository}
}
\value{
Returns the current package version
}
\description{
Get Remote (GitHub) Package Version
}
\examples{
get_remote_package_version("ScottishCovidResponse/SCRCdataAPI")

}
\keyword{internal}
