% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responsible_person.R
\name{get_responsible_person}
\alias{get_responsible_person}
\title{get_responsible_person}
\usage{
get_responsible_person(full_name, key)
}
\arguments{
\item{key}{}
}
\description{
get_responsible_person
}
