% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_startup_message.R
\name{get_startup_message}
\alias{get_startup_message}
\title{Get startup message}
\usage{
get_startup_message(repo, package)
}
\arguments{
\item{repo}{a \code{string} specifying the github username/repository}

\item{package}{a \code{string} specifying the package name}
}
\value{
returns the startup message
}
\description{
Get startup message
}
\examples{
\dontrun{
get_startup_message("ScottishCovidResponse/SCRCdata", "SCRCdata")
}

}
\keyword{internal}
