% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage_location.R
\name{get_storage_location}
\alias{get_storage_location}
\title{Get Storage Location}
\usage{
get_storage_location(location_id)
}
\arguments{
\item{location_id}{an \code{integer} specifying the id of an entry in the
storage_location table}
}
\value{
Returns a \code{list} of fields associated with the specified entry
}
\description{
Get storage location entry
}
\examples{
\dontrun{
get_storage_location(10404)
}

}
\keyword{internal}
