% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_table_optional}
\alias{get_table_optional}
\title{Get Optional Fields}
\usage{
get_table_optional(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a character vector of optional fields
}
\description{
Get Optional Fields
}
