% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_optional.R
\name{get_table_optional}
\alias{get_table_optional}
\title{Get optional fields}
\usage{
get_table_optional(table)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}
}
\value{
Returns a \code{data.frame} of optional fields and thier properties
}
\description{
Get optional fields
}
\keyword{internal}
