% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_optional.R
\name{get_table_optional}
\alias{get_table_optional}
\title{Get Optional Fields}
\usage{
get_table_optional(table, key)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{key}{API token from data.scrc.uk}
}
\value{
Returns a \code{data.frame} of optional fields and thier properties
}
\description{
Get Optional Fields
}
\keyword{internal}
