% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_optional.R
\name{get_table_optional}
\alias{get_table_optional}
\title{Get Optional Fields}
\usage{
get_table_optional(table, key, type = FALSE)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}
}
\value{
a dataframe of optional fieldsa and thier properties
}
\description{
Get Optional Fields
}
