% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_queryable.R
\name{get_table_queryable}
\alias{get_table_queryable}
\title{Get Queryable Fields}
\usage{
get_table_queryable(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a character vector of queryable fields
}
\description{
Get Queryable Fields
}
\keyword{internal}
