% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_queryable.R
\name{get_table_queryable}
\alias{get_table_queryable}
\title{Get queryable fields}
\usage{
get_table_queryable(table, endpoint)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
Returns a character vector of queryable fields
}
\description{
Get queryable fields
}
\keyword{internal}
