% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_table_readable}
\alias{get_table_readable}
\title{Get Readable Fields}
\usage{
get_table_readable(table, key)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}
}
\value{
a character vector of readable fields
}
\description{
Get Readable Fields
}
