% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_readable.R
\name{get_table_readable}
\alias{get_table_readable}
\title{Get Readable Fields}
\usage{
get_table_readable(table, key)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}
}
\value{
a dataframe of readable fields and their propertied
}
\description{
Get Readable Fields
}
