% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_readable.R
\name{get_table_readable}
\alias{get_table_readable}
\title{Get readable fields}
\usage{
get_table_readable(table, endpoint)
}
\arguments{
\item{table}{name of table}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
a dataframe of readable fields and their properties
}
\description{
Get readable fields
}
\keyword{internal}
