% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_required.R
\name{get_table_required}
\alias{get_table_required}
\title{Get Required Fields}
\usage{
get_table_required(table, key, type = FALSE)
}
\arguments{
\item{table}{name of table}

\item{key}{api key / token}
}
\value{
a dataframe of required fields and their properties
}
\description{
Get Required Fields
}
