% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_table_required}
\alias{get_table_required}
\title{Get Required Fields}
\usage{
get_table_required(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a character vector of required fields
}
\description{
Get Required Fields
}
