% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_required.R
\name{get_table_required}
\alias{get_table_required}
\title{Get required fields}
\usage{
get_table_required(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a dataframe of required fields and their properties
}
\description{
Get required fields
}
\keyword{internal}
