% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{get_table_writable}
\alias{get_table_writable}
\title{Get Writable Fields}
\usage{
get_table_writable(table)
}
\arguments{
\item{table}{name of table}
}
\value{
a character vector of writable fields
}
\description{
Get Writable Fields
}
