% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_writable.R
\name{get_table_writable}
\alias{get_table_writable}
\title{Get writable fields}
\usage{
get_table_writable(table)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}
}
\value{
Returns a character vector of writable fields
}
\description{
Get writable fields
}
\keyword{internal}
