% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Get tables from registry}
\usage{
get_tables(live = FALSE)
}
\arguments{
\item{live}{whether or not to get the tables directly from the API}
}
\value{
a character vector of tables
}
\description{
Use api endpoint to produce a list of tables
}
\keyword{internal}
