% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Get tables from registry}
\usage{
get_tables(endpoint = "http://localhost:8000/api/")
}
\arguments{
\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\value{
a character vector of tables
}
\description{
Use api endpoint to produce a list of tables
}
\keyword{internal}
