% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Get URL}
\usage{
get_url(table, query = list())
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{list} containing a valid query for the table, \emph{e.g.}
\code{list(field = value)}}
}
\value{
Returns a string or list of strings specifying the URL or URLs of
entries in a table
}
\description{
Retrieve URLs for particular entries or all entries in a table
}
\examples{
\dontrun{
# Retrieve URLs for all entries in a table
get_url(table = "storage_root")

# Retrieve the URL of a particular entry in a table
get_url(table = "storage_root", list(name = "github"))
}

}
