% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_numbers.R
\name{get_version_numbers}
\alias{get_version_numbers}
\title{get_version_numbers}
\usage{
get_version_numbers(name)
}
\arguments{
\item{a}{\code{string} specifying the name of the data product}
}
\value{

}
\description{
get_version_numbers
}
