% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_files.R
\name{github_files}
\alias{github_files}
\title{List files in GitHub repository}
\usage{
github_files(repo)
}
\arguments{
\item{repo}{a \code{string} specifying the github username/repository}
}
\description{
List files in GitHub repository
}
\keyword{internal}
