% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_intersection.R
\name{grid_intersection}
\alias{grid_intersection}
\title{Grid intersection}
\usage{
grid_intersection(shapefile, gridsize)
}
\arguments{
\item{shapefile}{shapefile}

\item{gridsize}{gridsize}
}
\description{
Grid intersection
}
\keyword{internal}
