% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_filename.R
\name{increment_filename}
\alias{increment_filename}
\title{increment_filename}
\usage{
increment_filename(path)
}
\arguments{
\item{path}{path}
}
\description{
Searches directory for duplicate files and increments filename.
}
