% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{initialise}
\usage{
initialise(config, script, endpoint)
}
\arguments{
\item{config}{a \code{string} specifying the location of the working
config file in the data store}

\item{script}{a \code{string} specifying the location of the submission
script in the data store}

\item{endpoint}{endpoint}
}
\description{
Reads in a working config file and generates new Code Run entry.
}
