% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{Initialise code run}
\usage{
initialise(config, script)
}
\arguments{
\item{config}{a \code{string} specifying the location of the working
config file in the data store}

\item{script}{a \code{string} specifying the location of the submission
script in the data store}
}
\value{
Returns an object of class \code{fdp, R6} containing metadata
required by the Data Pipeline API
}
\description{
Reads in a working config file, generates new Code Run entry, and returns
a handle containing various metadata.
}
