% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_current_version.R
\name{is_current_version}
\alias{is_current_version}
\title{Check if package is current version}
\usage{
is_current_version(
  repo = "ScottishCovidResponse/SCRCdataAPI",
  package = "SCRCdataAPI"
)
}
\arguments{
\item{repo}{repo name}

\item{package}{package name}
}
\value{
Returns true if the loaded package is the same version as the github
package
}
\description{
Check if package is current version
}
\keyword{internal}
