% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_queryable.R
\name{is_queryable}
\alias{is_queryable}
\title{is_queryable}
\usage{
is_queryable(table, query_parameter)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query_parameter}{a \code{string} or \code{vector} of field names}
}
\value{
Returns \code{TRUE} if the entry is queryable and \code{FALSE} if it
isn't
}
\description{
Check whether fields are queryable
}
\examples{
# Is "name" a queryable field in the "storage_root" table?
is_queryable("storage_root", "name")

# Are "not_a_field" and "name" queryable fields in the "storage_root" table?
is_queryable("storage_root", c("not_a_field", "name"))

# Is "name" a queryable field in the "not_a_table" table?
is_queryable("users", "name")

}
\keyword{internal}
