% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definitions.R
\name{is_queryable}
\alias{is_queryable}
\title{is queryable
Produces error \code{Unknown Table} if table does not exist}
\usage{
is_queryable(table, query_parameter)
}
\arguments{
\item{table}{name of table}

\item{query_parameter}{a string or vector of strings to check}
}
\value{
either true / false if a single string is provided or a vector or true or false if vector is provided
}
\description{
is queryable
Produces error \code{Unknown Table} if table does not exist
}
