% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_queryable.R
\name{is_queryable}
\alias{is_queryable}
\title{Check whether fields are queryable}
\usage{
is_queryable(table, query)
}
\arguments{
\item{table}{a \code{string} specifying the name of the table}

\item{query}{a \code{string} or \code{vector} of field names}
}
\value{
Returns \code{TRUE} if the entry is queryable and \code{FALSE} if it
isn't
}
\description{
Check whether fields are queryable
}
\keyword{internal}
