% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_component.R
\name{issue_with_component}
\alias{issue_with_component}
\title{issue_with_component}
\usage{
issue_with_component(
  index,
  handle,
  component,
  data_product,
  version,
  namespace,
  issue,
  severity
)
}
\arguments{
\item{index}{component_id}

\item{handle}{handle}

\item{component}{component}

\item{data_product}{data_product}

\item{version}{version}

\item{namespace}{namespace}

\item{issue}{issue}

\item{severity}{severity}
}
\description{
issue_with_component
}
