% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_component.R
\name{issue_with_component}
\alias{issue_with_component}
\title{issue_with_component}
\usage{
issue_with_component(
  issue,
  severity,
  data_product,
  namespace,
  component,
  version
)
}
\description{
issue_with_component
}
