% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_component.R
\name{issue_with_component}
\alias{issue_with_component}
\title{issue_with_component}
\usage{
issue_with_component(
  component_id,
  handle,
  component,
  data_product,
  version,
  issue,
  severity
)
}
\arguments{
\item{component_id}{component_id}

\item{handle}{handle}

\item{component}{component}

\item{data_product}{data_product}

\item{version}{version}

\item{issue}{issue}

\item{severity}{severity}
}
\value{

}
\description{
issue_with_component
}
