% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_dataproduct.R
\name{issue_with_dataproduct}
\alias{issue_with_dataproduct}
\title{issue_with_dataproduct}
\usage{
issue_with_dataproduct(issue, severity, data_product, namespace, version)
}
\description{
issue_with_dataproduct
}
