% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_dataproduct.R
\name{issue_with_dataproduct}
\alias{issue_with_dataproduct}
\title{issue_with_dataproduct}
\usage{
issue_with_dataproduct(
  index,
  handle,
  data_product,
  version,
  namespace,
  issue,
  severity
)
}
\arguments{
\item{index}{\code{numeric} value returned from
\code{create_array()}, referencing an input / output in the handle}

\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{version}{a \code{string} specifying the version of the data product}

\item{namespace}{a \code{string} specifying the namespace containing the
data product}

\item{issue}{a \code{string} describing the issue}

\item{severity}{a \code{numeric} value specifying the severity of the issue}
}
\description{
Function to attach an issue to data product in handle. To be written to the
data registry with \code{finalise()}.
}
\details{
An issue can be attached to a data product referenced by an \code{id} (returned
from \code{create_array()}) or
}
