% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_dataproduct.R
\name{issue_with_dataproduct}
\alias{issue_with_dataproduct}
\title{issue_with_dataproduct}
\usage{
issue_with_dataproduct(
  data_product_id,
  handle,
  data_product,
  version,
  issue,
  severity
)
}
\arguments{
\item{data_product_id}{\code{numeric} value returned from
\code{create_array()}, referencing an input / output in the handle}

\item{handle}{handle}

\item{data_product}{data_product}

\item{version}{version}

\item{issue}{issue}

\item{severity}{severity}
}
\description{
Function to attach an issue to data product in handle. To be written to the
data registry with \code{finalise()}.
}
\details{
An issue can be attached to a data product referenced by an \code{id} (returned
from \code{create_array()}) or
}
