% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_with_externalobject.R
\name{issue_with_externalobject}
\alias{issue_with_externalobject}
\title{issue_with_externalobject}
\usage{
issue_with_externalobject(description, severity, external_object_doi, version)
}
\arguments{
\item{description}{description}

\item{severity}{severity}

\item{external_object_doi}{external_object_doi}

\item{version}{version}
}
\description{
issue_with_externalobject
}
