% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_read.R
\name{link_read}
\alias{link_read}
\title{link_read}
\usage{
link_read(handle, alias)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{alias}{a \code{string} representing an external object in the
config.yaml file}
}
\description{
link_read
}
