% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_read.R
\name{link_read}
\alias{link_read}
\title{Link path to external format data}
\usage{
link_read(handle, data_product)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} representing an external object in the
config.yaml file}
}
\value{
Returns a \code{string} specifying the location of the data product
to be read
}
\description{
Link path to external format data
}
