% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_write.R
\name{link_write}
\alias{link_write}
\title{Link path for external format data}
\usage{
link_write(handle, data_product)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} representing an external object in the
config.yaml file}
}
\value{
Returns a \code{string} specifying the location in which the data
product should be written
}
\description{
Link path for external format data
}
