% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_author.R
\name{new_author}
\alias{new_author}
\title{Post entry to author table}
\usage{
new_author(
  family_name,
  given_name,
  identifier,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{family_name}{a \code{string} specifying the author's family name}

\item{given_name}{a \code{string} specifying the author's first name
and / or middle name and / or any initials}

\item{identifier}{(optional) a \code{string} specifying the full URL of
the \code{author}s identifier, \emph{e.g.} ORCID iD}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{author} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_author_org}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_organisation}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()}
}
\concept{new functions}
