% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_author.R
\name{new_author}
\alias{new_author}
\title{Post entry to author table}
\usage{
new_author(name, identifier, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{name}{a \code{string} specifying the full name or organisation name of
the \code{author}; note that at least one of name or identifier must be
specified}

\item{identifier}{(optional) a \code{string} specifying the full URL
identifier (\emph{e.g.} ORCiD or ROR ID) of the \code{author}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{author} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
