% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_code_repo_version.R
\name{new_code_repo_release}
\alias{new_code_repo_release}
\title{new_code_repo_release}
\usage{
new_code_repo_release(name, version, website, object, key)
}
\arguments{
\item{name}{e.g.}

\item{version}{e.g.}

\item{website}{e.g.}

\item{object}{e.g.}

\item{key}{key}
}
\description{
new_code_repo_release
}
