% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_code_repo_release.R
\name{new_code_repo_release}
\alias{new_code_repo_release}
\title{Post entry to code_repo_release table}
\usage{
new_code_repo_release(
  name,
  version,
  object_url,
  website,
  endpoint = "http://127.0.0.1:8000/api/"
)
}
\arguments{
\item{name}{a \code{string} specifying the name of an official release of
code}

\item{version}{a \code{string} specifying the version release
(conforming with semantic versioning syntax)}

\item{object_url}{a \code{string} specifying the URL of an \code{object}}

\item{website}{(optional) a \code{string} specifying the URL of the
website for this code release}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{code_repo_release} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
