% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coderun.R
\name{new_coderun}
\alias{new_coderun}
\title{new_coderun}
\usage{
new_coderun(
  run_date,
  run_identifier,
  code_repo_id = "",
  model_config = "",
  submission_script_id = "",
  inputs = list(),
  outputs = list(),
  key
)
}
\arguments{
\item{run_date}{e.g.}

\item{run_identifier}{e.g.}

\item{code_repo_id}{e.g.}

\item{model_config}{e.g.}

\item{submission_script_id}{e.g.}

\item{inputs}{e.g.}

\item{outputs}{e.g.}

\item{key}{key}
}
\description{
new_coderun
}
