% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product.R
\name{new_data_product}
\alias{new_data_product}
\title{new_data_product}
\usage{
new_data_product(name, description, responsible_person, type, versions, key)
}
\arguments{
\item{name}{Name of the data product}

\item{description}{Free text description of the data product}

\item{responsible_person}{Name of the responsible person}

\item{type}{Name of data product type}

\item{key}{GitHub key}
}
\description{
new_data_product
}
