% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_product.R
\name{new_data_product}
\alias{new_data_product}
\title{Post entry to data_product table}
\usage{
new_data_product(
  name,
  version,
  object_url,
  namespace_url,
  endpoint = "http://127.0.0.1:8000/api/"
)
}
\arguments{
\item{name}{a \code{string} specifying the name of the \code{data_product}}

\item{version}{a \code{string} specifying the version identifier of the
\code{data_product} (must conform to semantic versioning syntax)}

\item{object_url}{a \code{string} specifying the URL of the entry in the
\code{object} table}

\item{namespace_url}{a \code{string} specifying the URL of the entry in the
\code{namespace} table}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{data_product} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
