% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_external_object.R
\name{new_external_object}
\alias{new_external_object}
\title{new_external_object}
\usage{
new_external_object(
  doi_or_unique_name,
  primary_not_supplement,
  release_date,
  title,
  description,
  version,
  object,
  source,
  original_store
)
}
\arguments{
\item{original_store}{}
}
\description{
new_external_object
}
