% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_external_object.R
\name{new_external_object}
\alias{new_external_object}
\title{new_external_object}
\usage{
new_external_object(
  doi_or_unique_name,
  primary_not_supplement,
  release_date,
  title,
  description,
  version,
  object_id,
  source_id,
  original_store_id,
  key
)
}
\arguments{
\item{doi_or_unique_name}{e.g. "scottish coronavirus-covid-19-management-information"}

\item{primary_not_supplement}{e.g. TRUE}

\item{release_date}{e.g. "2010-07-11 12:15:00 BST"}

\item{title}{e.g. "scottish coronavirus-covid-19-management-information"}

\item{description}{e.g. "scottish coronavirus-covid-19-management-information dataset"}

\item{version}{e.g. "20100711.0"}

\item{object_id}{e.g. "https://data.scrc.uk/api/object/12/"}

\item{source_id}{e.g. "https://data.scrc.uk/api/source/9/"}

\item{original_store_id}{e.g. "https://data.scrc.uk/api/storage_location/19/"}

\item{key}{key}
}
\description{
new_external_object
}
