% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_external_object.R
\name{new_external_object}
\alias{new_external_object}
\title{new_external_object}
\usage{
new_external_object(
  doi_or_unique_name,
  primary_not_supplement,
  release_date,
  title,
  description,
  version,
  object_id,
  source_id,
  original_store_id,
  key
)
}
\arguments{
\item{doi_or_unique_name}{e.g.}

\item{primary_not_supplement}{e.g.}

\item{release_date}{e.g.}

\item{title}{e.g.}

\item{description}{e.g.}

\item{version}{e.g.}

\item{object_id}{e.g.}

\item{source_id}{e.g.}

\item{original_store_id}{e.g.}

\item{key}{key}
}
\description{
new_external_object
}
