% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_issue.R
\name{new_issue}
\alias{new_issue}
\title{new_issue}
\usage{
new_issue(severity, description, object_issues, component_issues, key)
}
\arguments{
\item{severity}{an \code{integer} specifying the severity of the \code{issue} \emph{e.g.} 16}

\item{description}{a \code{string} containing a free text description of the
\code{issue} \emph{e.g.} "Data dump caused a spike on the 15th of June"}

\item{object_issues}{a \code{list} of \code{object} URLs with which the \code{issue} is
associated \emph{e.g.} list("https://data.scrc.uk/api/object/152/")}

\item{component_issues}{a \code{list} of \code{object_component} URLs with which the
\code{issue} is associated
\emph{e.g.} list("https://data.scrc.uk/api/object_component/902/")}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{issue} table in the data registry
}
