% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_issue.R
\name{new_issue}
\alias{new_issue}
\title{Post entry to issue table}
\usage{
new_issue(
  severity,
  description,
  component_issues,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{severity}{an \code{integer} specifying the severity of the \code{issue}}

\item{description}{a \code{string} containing a free text description of the
\code{issue}}

\item{component_issues}{a \code{list} of \code{object_component} URLs with
which the \code{issue} is associated; this can be an empty list}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{issue} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
