% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_keyword.R
\name{new_keyword}
\alias{new_keyword}
\title{new_keyword}
\usage{
new_keyword(keyphrase, object_id, key)
}
\arguments{
\item{keyphrase}{e.g.}

\item{object_id}{e.g.}

\item{key}{key}
}
\description{
new_keyword
}
