% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_keyword.R
\name{new_keyword}
\alias{new_keyword}
\title{new_keyword}
\usage{
new_keyword(keyphrase, object_id, key)
}
\arguments{
\item{keyphrase}{a \code{string} specifying a single keyword or keyphrase \emph{e.g.}
"covid-19" or "cardiovascular diseases"}

\item{object_id}{a \code{string} specifying the API URL of the
associated \code{object} table \emph{e.g.} "https://data.scrc.uk/api/object/31817/"}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{keyword} table in the data registry
}
