% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_keyword.R
\name{new_keyword}
\alias{new_keyword}
\title{Post entry to keyword table}
\usage{
new_keyword(
  object_url,
  keyphrase,
  identifier,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{object_url}{a \code{string} specifying the URL of an \code{object}}

\item{keyphrase}{a \code{string} a \code{string} containing a free text
key phrase}

\item{identifier}{(optional) a \code{string} specifying the URL of ontology
annotation to associate with this \code{keyword}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{keyword} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
