% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_namespace.R
\name{new_namespace}
\alias{new_namespace}
\title{Post entry to namespace table}
\usage{
new_namespace(name)
}
\arguments{
\item{name}{a \code{string} specifying the name of the namespace \emph{e.g.} "SCRC"}
}
\description{
Upload information to the \code{namespace} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
