% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object.R
\name{new_object}
\alias{new_object}
\title{Post entry to object table}
\usage{
new_object(storage_location_uri, description = "", issues)
}
\arguments{
\item{storage_location_uri}{a \code{string} specifying the URI of an entry
in the \code{storage_location} table}

\item{description}{(optional) a \code{string} containing a free text
description of the \code{object}}

\item{issues}{a \code{list} of Issues URLs to associate with this Object

Note that the \code{object} table contains \code{issues} as an
additional optional field. This is not included here. Instead use
\code{attach_issue()} and associated functionality to attach issues to
objects and object components.}
}
\description{
Upload information to the \code{object} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
