% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object.R
\name{new_object}
\alias{new_object}
\title{Post to object table}
\usage{
new_object(storage_location_id, description = "", key)
}
\arguments{
\item{storage_location_id}{a \code{string} specifying the API URL of the
associated \code{storage_location} table
\emph{e.g.} "https://data.scrc.uk/api/storage_location/411/"}

\item{description}{(optional) a \code{string} containing a free text
description of the \code{object_component}}

\item{key}{API token from data.scrc.uk

Note that the \code{object} table contains \code{issues} as an
additional optional field. This is not included here. Instead use
\code{attach_issue()} and associated functionality to attach issues to
objects and object components.}
}
\description{
Upload information to the \code{object} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
