% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object_component.R
\name{new_object_component}
\alias{new_object_component}
\title{Post to object_component table}
\usage{
new_object_component(name, object_id, description = "", key)
}
\arguments{
\item{name}{a \code{string} specifying the name of the
\code{object_component}, unique in the context of \code{object_component}
and its \code{object} reference}

\item{object_id}{a \code{string} specifying the API URL of the
associated \code{object} table}

\item{description}{(optional) a \code{string} containing a free text
description of the \code{object_component}}

\item{key}{API token from data.scrc.uk

Note that the \code{object_component} table contains \code{issues} is an
additional optional field. This is not included here. Instead use
\code{attach_issue()} and associated functionality to attach issues to
objects and objet components.}
}
\description{
Upload information to the \code{object_component} table in the data registry
}
\examples{
\dontrun{
\donttest{
new_object_component(name = "age_group/week-persons-country-all_deaths",
                     object_id = "https://data.scrc.uk/api/object/156/",
                     description = "A very useful component",
                     key = key)
}}

}
