% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_object_component.R
\name{new_object_component}
\alias{new_object_component}
\title{Post entry to object_component table}
\usage{
new_object_component(name, object_id, description = "")
}
\arguments{
\item{name}{a \code{string} specifying the name of the
\code{object_component}, unique in the context of \code{object_component}
and its \code{object} reference}

\item{object_id}{a \code{string} specifying the API URL of the
associated \code{object} table}

\item{description}{(optional) a \code{string} containing a free text
description of the \code{object_component}

Note that the \code{object_component} table contains \code{issues} as an
additional optional field. This is not included here. Instead use
\code{attach_issue()} and associated functionality to attach issues to
objects and objet components.}
}
\description{
Upload information to the \code{object_component} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
