% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_processing_script.R
\name{new_processing_script}
\alias{new_processing_script}
\title{new_processing_script}
\usage{
new_processing_script(name, responsible_person, store, versions, key)
}
\arguments{
\item{name}{Name of the processing script}

\item{responsible_person}{Name of the responsible person}

\item{store}{Name of the storage_location, e.g. "Model File"}

\item{key}{GitHub key}
}
\description{
new_processing_script
}
