% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_processing_script_version.R
\name{new_processing_script_version}
\alias{new_processing_script_version}
\title{new_processing_script_version}
\usage{
new_processing_script_version(
  version_identifier,
  responsible_person,
  supercedes,
  processing_script,
  store,
  accessibility,
  data_product_versions,
  key
)
}
\arguments{
\item{key}{GitHub key}
}
\description{
new_processing_script_version
}
