% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_quality_controlled.R
\name{new_quality_controlled}
\alias{new_quality_controlled}
\title{Post entry to quality_controlled table}
\usage{
new_quality_controlled(object_url, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{object_url}{a \code{string} specifying the URL of an \code{object}}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{quality_controlled} table in the data
registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
