% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{new_source}
\usage{
new_source(name, abbreviation, website, key)
}
\arguments{
\item{name}{e.g. "Scottish Government Open Data Repository"}

\item{abbreviation}{e.g. "Scottish Government Open Data Repository"}

\item{website}{e.g. "https://statistics.gov.scot/"}

\item{key}{key}
}
\description{
new_source
}
