% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{Post to source table}
\usage{
new_source(name, abbreviation, website = "", key)
}
\arguments{
\item{name}{a \code{string} specifying the name of the source
\emph{e.g.} "Scottish Government Open Data Repository"}

\item{abbreviation}{a \code{string} specifying the common abbreviation
of the source (if available) \emph{e.g.} "Scottish Government Open Data Repository"}

\item{website}{(optional) a \code{string} specifying the website URL
associated with the data source \emph{e.g.} "https://statistics.gov.scot/"}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{source} table in the data registry
}
