% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{new_source}
\usage{
new_source(name, description, responsible_person, store, source_type, key)
}
\arguments{
\item{name}{Name of the source}

\item{description}{Free text description of the source}

\item{responsible_person}{Name of the responsible person}

\item{store}{Name of the storage_location, e.g. "Model File"}

\item{source_type}{Name of source type}

\item{key}{GitHub key}
}
\description{
new_source
}
