% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{Post entry to source table}
\usage{
new_source(name, abbreviation, website = "")
}
\arguments{
\item{name}{a \code{string} specifying the name of the source
\emph{e.g.} "Scottish Government Open Data Repository"}

\item{abbreviation}{a \code{string} specifying the common abbreviation
of the source (if available) \emph{e.g.} "Scottish Government Open Data Repository"}

\item{website}{(optional) a \code{string} specifying the website URL
associated with the data source \emph{e.g.} "https://statistics.gov.scot/"}
}
\description{
Upload information to the \code{source} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
