% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source_type.R
\name{new_source_type}
\alias{new_source_type}
\title{new_source_type}
\usage{
new_source_type(name, description)
}
\arguments{
\item{name}{name of the source}

\item{description}{free text description of the source}
}
\description{
new_source_type
}
