% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_location.R
\name{new_storage_location}
\alias{new_storage_location}
\title{new_storage_location}
\usage{
new_storage_location(id, uri, text, hash, storage_root, accessibility_id)
}
\arguments{
\item{accessibility_id}{}
}
\description{
new_storage_location
}
