% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_location.R
\name{new_storage_location}
\alias{new_storage_location}
\title{new_storage_location}
\usage{
new_storage_location(path, hash, storage_root_id, key)
}
\arguments{
\item{path}{e.g.}

\item{hash}{e.g.}

\item{storage_root_id}{e.g.}

\item{key}{key}
}
\description{
new_storage_location
}
