% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_location.R
\name{new_storage_location}
\alias{new_storage_location}
\title{new_storage_location}
\usage{
new_storage_location(path, hash, storage_root_id, key)
}
\arguments{
\item{path}{a \code{string} specifying the path from the \code{storage_root}
URI to the item location, which when appended to \code{storage_root} URI
produces a complete URL}

\item{hash}{a \code{string} specifying the SHA1 hash of the file stored in
\code{storage_location}}

\item{storage_root_id}{a \code{string} specifying the API URL of the
associated \code{storage_root} table}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{storage_location} table in the data registry
}
\examples{
\dontrun{
\donttest{
new_storage_location(path = "master/SCRC/human/infection/SARS-CoV-2/latent-period/0.1.0.toml",
hash = get_hash(file.path("local", "file", "location", "0.1.0.toml")),
storage_root_id = "https://data.scrc.uk/api/storage_root/14/",
key)
}}

}
