% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_location.R
\name{new_storage_location}
\alias{new_storage_location}
\title{Post entry to storage_location table}
\usage{
new_storage_location(
  path,
  hash,
  public,
  storage_root_url,
  endpoint = "http://127.0.0.1:8000/api/"
)
}
\arguments{
\item{path}{a \code{string} specifying the path from the \code{storage_root}
URI to the item location, which when appended to \code{storage_root} URI
produces a complete URL}

\item{hash}{a \code{string} specifying the SHA1 hash of a file stored in
\code{storage_location}}

\item{public}{a \code{boolean} indicating whether the \code{storage_location}
is public or not (default is \code{TRUE})}

\item{storage_root_url}{a \code{string} specifying the URL of an entry in
the \code{storage_root} table}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{storage_location} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_root}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
