% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_root.R
\name{new_storage_root}
\alias{new_storage_root}
\title{Post entry to storage_root table}
\usage{
new_storage_root(name, root, accessibility = 0, key)
}
\arguments{
\item{name}{a \code{string} specifying the name of the \code{storage_root}
\emph{e.g.} "boydorr"}

\item{root}{a \code{string} specifying the URI to the root of a
\code{storage_location}, which is then prepended to a \code{storage_location}
\emph{e.g.} "ftp://boydorr.gla.ac.uk/scrc/"}

\item{accessibility}{(optional) an \code{integer} value for the accessibility
enum, where 0 is public (default) and 1 is private}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload information to the \code{storage_root} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_text_file}()}
}
\concept{new functions}
