% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_root.R
\name{new_storage_root}
\alias{new_storage_root}
\title{new_storage_root}
\usage{
new_storage_root(name, description, uri, type, key)
}
\arguments{
\item{name}{Name of the storage root, e.g. "Boydorr"}

\item{description}{Free text description of the storage root, e.g. "Boydorr
FTP server"}

\item{uri}{Link to the root of the storage, e.g. "ftp://srv/ftp/scrc"}

\item{type}{Name of storage type, e.g. ftp, https}

\item{key}{GitHub key}
}
\description{
new_storage_root
}
