% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_root.R
\name{new_storage_root}
\alias{new_storage_root}
\title{new_storage_root}
\usage{
new_storage_root(name, description, uri, type)
}
\arguments{
\item{name}{name of the storage root, e.g. "Boydorr"}

\item{description}{free text description of the storage root, e.g. "Boydorr
FTP server"}

\item{uri}{link to the root of the storage, e.g. "ftp://srv/ftp/scrc"}

\item{type}{reference to the storage type used}
}
\description{
new_storage_root
}
