% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_storage_root.R
\name{new_storage_root}
\alias{new_storage_root}
\title{Post entry to storage_root table}
\usage{
new_storage_root(root, local, endpoint = "http://127.0.0.1:8000/api/")
}
\arguments{
\item{root}{a \code{string} specifying the URI of a
\code{storage_location}, which when prepended to a \code{storage_location}
produces a complete URI to a file}

\item{local}{(optional) a \code{boolean} indicating whether the
\code{storage_root} is local or not}

\item{endpoint}{a \code{string} specifying the registry endpoint}
}
\description{
Upload information to the \code{storage_root} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_code_run}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_file_type}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_licence}()},
\code{\link{new_namespace}()},
\code{\link{new_object}()},
\code{\link{new_object_component}()},
\code{\link{new_quality_controlled}()},
\code{\link{new_storage_location}()},
\code{\link{new_user_author}()}
}
\concept{new functions}
\keyword{internal}
