% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_text_file.R
\name{new_text_file}
\alias{new_text_file}
\title{Post entry to text_file table}
\usage{
new_text_file(text)
}
\arguments{
\item{text}{a \code{string} containing a free text \emph{i.e.} scripts that are
only a few lines long and don't have a home elsewhere}
}
\description{
Upload information to the \code{text_file} table in the data registry
}
\seealso{
Other new functions: 
\code{\link{new_author}()},
\code{\link{new_code_repo_release}()},
\code{\link{new_coderun}()},
\code{\link{new_data_product}()},
\code{\link{new_external_object}()},
\code{\link{new_issue}()},
\code{\link{new_keyword}()},
\code{\link{new_namespace}()},
\code{\link{new_object_component}()},
\code{\link{new_object}()},
\code{\link{new_source}()},
\code{\link{new_storage_location}()},
\code{\link{new_storage_root}()}
}
\concept{new functions}
