% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_source_data.R
\name{post_source_data}
\alias{post_source_data}
\title{post_source_data}
\usage{
post_source_data(
  storage_type,
  storage_root,
  storage_location,
  accessibility,
  source_type,
  source,
  source_version_identifier,
  source_version_description,
  responsible_person,
  source_version_supercedes,
  key
)
}
\arguments{
\item{storage_type}{Name of storage type, e.g. ftp, https}

\item{storage_root}{Name of the storage root, e.g. "Boydorr"}

\item{storage_location}{Name of the storage_location, e.g. "Model File"}

\item{accessibility}{Name of accessibility type}

\item{source_type}{Name of the source type}

\item{source}{Name of the source}

\item{source_version_description}{Free text description of the source_version}

\item{responsible_person}{Name of the responsible person}

\item{key}{GitHub key}
}
\description{
post_source_data
}
