% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue.R
\name{raise_issue}
\alias{raise_issue}
\title{raise_issue}
\usage{
raise_issue(
  index,
  handle,
  component = NA,
  data_product,
  version,
  namespace,
  issue,
  severity,
  whole_object
)
}
\arguments{
\item{index}{component_id}

\item{handle}{handle}

\item{component}{component}

\item{data_product}{data_product}

\item{version}{version}

\item{namespace}{namespace}

\item{issue}{issue}

\item{severity}{severity}

\item{whole_object}{whole_object}
}
\description{
raise_issue
}
