% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue_config.R
\name{raise_issue_config}
\alias{raise_issue_config}
\title{raise_issue_config}
\usage{
raise_issue_config(handle, issue, severity)
}
\arguments{
\item{handle}{handle}

\item{issue}{issue}

\item{severity}{severity}
}
\description{
raise_issue_config
}
