% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_issue_repo.R
\name{raise_issue_repo}
\alias{raise_issue_repo}
\title{Raise issue with remote repository}
\usage{
raise_issue_repo(handle, issue, severity)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{issue}{a \code{string} specifying the issue}

\item{severity}{a \code{numeric} value specifying the severity}
}
\description{
Raise issue with remote repository
}
