% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDataPipeline-package.R
\docType{package}
\name{rDataPipeline-package}
\alias{rDataPipeline-package}
\alias{rDataPipeline}
\title{rDataPipeline}
\description{
FAIR Data Pipeline API
}
\details{
For more information see
\url{https://www.fairdatapipeline.org/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.fairdatapipeline.org/rDataPipeline/}
  \item \url{https://github.com/FAIRDataPipeline/rDataPipeline}
  \item Report bugs at \url{https://github.com/FAIRDataPipeline/rDataPipeline/issues}
}

}
\author{
\strong{Maintainer}: Ryan Field \email{ryan.field@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-4424-9890}{ORCID})

Authors:
\itemize{
  \item Sonia Mitchell (\href{https://orcid.org/0000-0003-1536-2066}{ORCID})
}

}
