% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_array.R
\name{read_array}
\alias{read_array}
\title{Read component from array-type H5 file}
\usage{
read_array(filepath, component)
}
\arguments{
\item{filepath}{a \code{string} specifying the path and filename of the file to
be read}

\item{component}{a \code{string} specifying a location within the hdf5 file,
e.g. "location/per_week/all_deaths"}
}
\value{
Returns an array with attached \code{Dimension_i_title},
\code{Dimension_i_units}, \code{Dimension_i_values}, and \code{units}
attributes, if available
}
\description{
Function to read array type data from hdf5 file.
}
