% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_array.R
\name{read_array}
\alias{read_array}
\title{Read array component from HDF5 file}
\usage{
read_array(handle, data_product, component)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying a data product}

\item{component}{a \code{string} specifying a data product component}
}
\value{
Returns an array with attached \code{Dimension_i_title},
\code{Dimension_i_units}, \code{Dimension_i_values}, and \code{units}
attributes, if available
}
\description{
Function to read array type data from hdf5 file.
}
