% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_distribution.R
\name{read_distribution}
\alias{read_distribution}
\title{read_distribution}
\usage{
read_distribution(filename)
}
\arguments{
\item{filename}{a \code{string} specifying the name of the toml file, e.g.
"0.1.0.toml"}
}
\description{
read_distribution
}
