% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estimate.R
\name{read_estimate}
\alias{read_estimate}
\title{Read estimate from TOML file}
\usage{
read_estimate(handle, data_product, component)
}
\arguments{
\item{handle}{\code{fdp} object}

\item{data_product}{a \code{string} specifying a data product}

\item{component}{a \code{string} specifying a data product component}
}
\description{
Function to read point-estimate type data from toml file.
}
