% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estimate.R
\name{read_estimate}
\alias{read_estimate}
\title{Read estimate from TOML file}
\usage{
read_estimate(filepath)
}
\arguments{
\item{filepath}{a \code{string} specifying the path and filename of the
file to be read}
}
\description{
Function to read point-estimate type data from toml file.
}
