% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estimate.R
\name{read_estimate}
\alias{read_estimate}
\title{Read estimate component from TOML file}
\usage{
read_estimate(handle, data_product, component)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying a data product}

\item{component}{a \code{string} specifying a data product component}
}
\description{
Function to read point-estimate type data from toml file.
}
