% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table}
\alias{read_table}
\title{Read table component from HDF5 file}
\usage{
read_table(handle, data_product, component)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying a data product}

\item{component}{a \code{string} specifying a data product component}
}
\value{
Returns a \code{data.frame} with attached \code{column_units}
attributes, if available
}
\description{
Function to read table type data from hdf5 file.
}
