% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_data_product.R
\name{register_data_product}
\alias{register_data_product}
\title{register_data_product}
\usage{
register_data_product(this_metadata, handle, datastore)
}
\arguments{
\item{this_metadata}{text}

\item{handle}{text}

\item{datastore}{text}
}
\description{
register_data_product
}
