% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_external_object.R
\name{register_external_object}
\alias{register_external_object}
\title{register_external_object}
\usage{
register_external_object(register_this, datastore, namespace, filename)
}
\arguments{
\item{register_this}{metadata}

\item{datastore}{a \code{string} specifying the default local data storage
location}

\item{namespace}{a \code{string} specifying the name of the namespace}

\item{filename}{a \code{string} specifying the filename}
}
\description{
register_external_object
}
