% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_component.R
\name{register_issue_component}
\alias{register_issue_component}
\title{register_issue_component}
\usage{
register_issue_component(
  issue,
  severity,
  data_product,
  namespace,
  component,
  version
)
}
\arguments{
\item{issue}{issue}

\item{severity}{severity}

\item{data_product}{data_product}

\item{namespace}{namespace}

\item{component}{component}

\item{version}{version}
}
\description{
register_issue_component
}
