% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_issue_dataproduct.R
\name{register_issue_dataproduct}
\alias{register_issue_dataproduct}
\title{register_issue_dataproduct}
\usage{
register_issue_dataproduct(issue, severity, data_product, namespace, version)
}
\arguments{
\item{issue}{issue}

\item{severity}{severity}

\item{data_product}{data_product}

\item{namespace}{namespace}

\item{version}{version}
}
\description{
register_issue_dataproduct
}
