% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_read.R
\name{resolve_read}
\alias{resolve_read}
\title{resolve_read}
\usage{
resolve_read(handle, data_product, component = NA)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{component}{a \code{string} specifying the name of data product
component}
}
\description{
resolve_read
}
