% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_version.R
\name{resolve_version}
\alias{resolve_version}
\title{resolve_version}
\usage{
resolve_version(version, data_product, namespace_id)
}
\arguments{
\item{version}{version number}

\item{data_product}{data_product}

\item{namespace_id}{namespace_id}
}
\description{
resolve_version
}
\keyword{internal}
