% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_write.R
\name{resolve_write}
\alias{resolve_write}
\title{resolve_data_product}
\usage{
resolve_write(handle, data_product, file_type)
}
\arguments{
\item{handle}{an object of class \code{fdp, R6} containing metadata required
by the Data Pipeline API}

\item{data_product}{a \code{string} specifying the name of the data product}

\item{file_type}{(optional) a \code{string} specifying the file type; when
missing, \code{file_type} will be read from the config file}
}
\description{
resolve_data_product
}
\keyword{internal}
