% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_product.R
\name{upload_data_product}
\alias{upload_data_product}
\title{upload_data_product}
\usage{
upload_data_product(
  storage_root_id,
  path,
  name = name,
  component_name,
  filename,
  version,
  namespace_id,
  key
)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{path}{e.g.}

\item{name}{e.g.}

\item{component_name}{e.g.}

\item{filename}{e.g.}

\item{version}{e.g.}

\item{namespace_id}{e.g.}

\item{key}{key}
}
\description{
upload_data_product
}
