% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_product.R
\name{upload_data_product}
\alias{upload_data_product}
\title{Upload data_product metadata to the data registry}
\usage{
upload_data_product(
  storage_root_id,
  name,
  component_name,
  processed_path,
  product_path,
  version,
  namespace_id,
  key
)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{name}{e.g.}

\item{component_name}{(optional) used for toml files, not hdf5}

\item{processed_path}{e.g.}

\item{product_path}{e.g.}

\item{version}{e.g.}

\item{namespace_id}{e.g.}

\item{key}{API token from data.scrc.uk}
}
\description{
Upload data_product metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_github_repo}()},
\code{\link{upload_object_links}()},
\code{\link{upload_paper}()},
\code{\link{upload_source_data}()},
\code{\link{upload_submission_script}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
