% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_github_repo.R
\name{upload_github_repo}
\alias{upload_github_repo}
\title{Post github_repo metadata to the data registry}
\usage{
upload_github_repo(storage_root_id, repo, hash, version)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{repo}{repo}

\item{hash}{e.g.}

\item{version}{version}
}
\description{
Post github_repo metadata to the data registry
}
\concept{upload functions}
