% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_github_repo.R
\name{upload_github_repo}
\alias{upload_github_repo}
\title{Post github_repo metadata to the data registry}
\usage{
upload_github_repo(storage_root_id, repo, hash, version, key)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{repo}{repo}

\item{hash}{e.g.}

\item{version}{version}

\item{key}{API token from data.scrc.uk}
}
\description{
Post github_repo metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_data_product}()},
\code{\link{upload_object_links}()},
\code{\link{upload_paper}()},
\code{\link{upload_source_data}()},
\code{\link{upload_submission_script}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
