% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_github_repo.R
\name{upload_github_repo}
\alias{upload_github_repo}
\title{upload_github_repo}
\usage{
upload_github_repo(storage_root_id, repo, hash, version, key)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{repo}{repo}

\item{hash}{e.g.}

\item{version}{version}

\item{key}{API token from data.scrc.uk}
}
\description{
upload_github_repo
}
