% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_object_links.R
\name{upload_object_links}
\alias{upload_object_links}
\title{upload_object_links}
\usage{
upload_object_links(
  run_date,
  run_identifier,
  code_repo_id,
  submission_script_id,
  inputs = list(),
  outputs = list(),
  key = key
)
}
\arguments{
\item{run_date}{e.g.}

\item{run_identifier}{e.g.}

\item{code_repo_id}{e.g.}

\item{submission_script_id}{e.g.}

\item{inputs}{e.g.}

\item{outputs}{e.g.}

\item{key}{key}
}
\description{
upload_object_links
}
