% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_object_links.R
\name{upload_object_links}
\alias{upload_object_links}
\title{Post object_links metadata to the data registry}
\usage{
upload_object_links(
  run_date,
  description,
  code_repo_id,
  submission_script_id,
  inputs = list(),
  outputs = list()
)
}
\arguments{
\item{run_date}{e.g.}

\item{description}{e.g.}

\item{code_repo_id}{e.g.}

\item{submission_script_id}{e.g.}

\item{inputs}{e.g.}

\item{outputs}{e.g.}
}
\description{
Post object_links metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_github_repo}()},
\code{\link{upload_paper}()},
\code{\link{upload_source_data}()},
\code{\link{upload_submission_script}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
