% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_paper.R
\name{upload_paper}
\alias{upload_paper}
\title{upload_paper}
\usage{
upload_paper(
  title,
  authors,
  journal,
  journal_abbreviation,
  journal_website,
  release_date,
  abstract,
  keywords,
  doi,
  primary_not_supplement = TRUE,
  version = 1,
  key
)
}
\arguments{
\item{title}{a \code{string} specifying the title of the paper}

\item{authors}{a \code{string} specifying the authors}

\item{journal}{a \code{string} specifying the full journal name}

\item{journal_abbreviation}{a \code{string} specifying the journal
abbreviation}

\item{journal_website}{a \code{string} specifying the journal homepage}

\item{release_date}{a \code{POSIXct} of format "\%Y-\%m-\%d \%H:\%M:\%S"
specifying the release date of the paper}

\item{abstract}{a \code{string} specifying the abstract}

\item{keywords}{a \code{string} specifying keywords or keyphrases
seperated by "and", e.g. "keyword1 and keyword2 and key phrase1 and keyword3"}

\item{doi}{a \code{string} specifying the doi}

\item{primary_not_supplement}{(optional) an object of class \code{logical}
where \code{TRUE} is the primary source (default) and false is a supplement}

\item{version}{(optional) a \code{string} specifying the version number}

\item{key}{key}
}
\description{
upload_paper
}
\examples{
title <- "Covid-19: A systemic disease treated with a wide-ranging approach: A case report"
authors <- "Massabeti, Rosanna and Cipriani, Maria Stella and Valenti, Ivana"
journal <- "Journal of Population Therapeutics and Clinical Pharmacology"
journal_abbreviation <- "J Popul Ther Clin Pharmacol"
journal_website <- "https://www.jptcp.com/index.php/jptcp"
release_date <- as.POSIXct("2020-01-01 12:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")
abstract <- "At the end of December 2019, the Health Commission of the"
keywords <- "covid-19 and coronavirus disease and monoclonal antibodies and non-invasive mechanical ventilation and treatment"
doi <- "10.15586/jptcp.v27iSP1.691"

}
