% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_source_data.R
\name{upload_source_data}
\alias{upload_source_data}
\title{upload_source_data}
\usage{
upload_source_data(
  dataset,
  source,
  source_root,
  source_path,
  local_path,
  storage_root,
  target_path,
  download_date,
  version,
  key
)
}
\arguments{
\item{dataset}{e.g. "scottish deaths-involving-coronavirus-covid-19"}

\item{source}{e.g. "Scottish Government Open Data Repository"}

\item{source_root}{e.g. "https://statistics.gov.scot/sparql.json?query="}

\item{source_path}{e.g. some query}

\item{local_path}{e.g. "data-raw/deaths-involving-coronavirus-covid-19.csv"}

\item{storage_root}{e.g. "boydorr"}

\item{target_path}{e.g. "human/infection/SARS-CoV-2/scotland/mortality/deaths-involving-coronavirus-covid-19.csv"}

\item{download_date}{e.g. as.Date("2010-07-09", "\%Y-\%m-\%d")}

\item{version}{e.g. 0}

\item{key}{}

\item{accessibility}{e.g. "public"}
}
\description{
upload_source_data
}
