% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_source_data.R
\name{upload_source_data}
\alias{upload_source_data}
\title{Post source_data metadata to the data registry}
\usage{
upload_source_data(
  doi_or_unique_name,
  original_source_id,
  original_root_id,
  original_path,
  primary_not_supplement = TRUE,
  local_path,
  storage_root_id,
  target_path,
  download_date,
  version
)
}
\arguments{
\item{doi_or_unique_name}{e.g. "scottish deaths-involving-coronavirus-covid-19"}

\item{original_source_id}{e.g. "https://data.scrc.uk/api/source/9/"}

\item{original_root_id}{e.g.}

\item{original_path}{e.g. query}

\item{primary_not_supplement}{e.g.}

\item{local_path}{e.g. "data-raw/human/infection/SARS-CoV-2/scotland/cases_and_management/20100711.0.csv"}

\item{storage_root_id}{e.g. "https://data.scrc.uk/api/storage_root/9/"}

\item{target_path}{e.g. "human/infection/SARS-CoV-2/scotland/cases_and_management/20100711.0.csv"}

\item{download_date}{e.g. as.POSIXct("2010-07-11 12:15:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")}

\item{version}{e.g. "20100711.0"}
}
\description{
Post source_data metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_data_product}()},
\code{\link{upload_github_repo}()},
\code{\link{upload_object_links}()},
\code{\link{upload_paper}()},
\code{\link{upload_submission_script}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
