% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_source_data.R
\name{upload_source_data}
\alias{upload_source_data}
\title{upload_source_data}
\usage{
upload_source_data(
  doi_or_unique_name,
  original_source_id,
  original_root_id,
  original_path,
  local_path,
  storage_root_id,
  target_path,
  download_date,
  version,
  key
)
}
\arguments{
\item{doi_or_unique_name}{e.g. "scottish deaths-involving-coronavirus-covid-19"}

\item{original_source_id}{e.g. "https://data.scrc.uk/api/source/9/"}

\item{original_root_id}{e.g.}

\item{original_path}{e.g. query}

\item{local_path}{e.g. "data-raw/human/infection/SARS-CoV-2/scotland/cases_and_management/20100711.0.csv"}

\item{storage_root_id}{e.g. "https://data.scrc.uk/api/storage_root/9/"}

\item{target_path}{e.g. "human/infection/SARS-CoV-2/scotland/cases_and_management/20100711.0.csv"}

\item{download_date}{e.g. as.POSIXct("2010-07-11 12:15:00", format = "\%Y-\%m-\%d \%H:\%M:\%S")}

\item{version}{e.g. "20100711.0"}

\item{key}{API token from data.scrc.uk}
}
\description{
upload_source_data
}
