% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_submission_script.R
\name{upload_submission_script}
\alias{upload_submission_script}
\title{Upload submission_script metadata to the data registry}
\usage{
upload_submission_script(storage_root_id, path, hash, text, run_date)
}
\arguments{
\item{storage_root_id}{e.g.}

\item{path}{e.g. "ScottishCovidResponse/SCRCdata"}

\item{hash}{e.g.}

\item{text}{text}

\item{run_date}{e.g.}
}
\description{
Upload submission_script metadata to the data registry
}
\seealso{
Other upload functions: 
\code{\link{upload_data_product}()},
\code{\link{upload_github_repo}()},
\code{\link{upload_object_links}()},
\code{\link{upload_paper}()},
\code{\link{upload_source_data}()},
\code{\link{upload_toml_to_github}()}
}
\concept{upload functions}
