% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fields.R
\name{validate_fields}
\alias{validate_fields}
\title{validate_fields}
\usage{
validate_fields(table, data, key, online = TRUE)
}
\arguments{
\item{table}{table name as character}

\item{data}{data as a named list}

\item{key}{API Token as character}

\item{online}{\code{boolean} if FALSE will not check if data registry objects exist}
}
\description{
Function to validate post data
}
