% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fields.R
\name{validate_fields}
\alias{validate_fields}
\title{Validate fields}
\usage{
validate_fields(table, data, endpoint)
}
\arguments{
\item{table}{table name as character}

\item{data}{data as a named list}

\item{endpoint}{endpoint}
}
\value{
Returns
}
\description{
Function to validate fields in post data
}
