% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_post_data.R
\name{validate_post_data}
\alias{validate_post_data}
\title{validate_post_data
Function to validate post data}
\usage{
validate_post_data(table, data, key)
}
\arguments{
\item{table}{table name as character}

\item{data}{data as a named list}

\item{key}{API Token as character}
}
\description{
validate_post_data
Function to validate post data
}
